Network.AWS.SNS.AddPermission

Description

Adds a statement to a topic's access control policy, granting access for the specified AWS accounts to the specified actions.

http://docs.aws.amazon.com/sns/latest/api/API_AddPermission.html

Request

data AddPermission

Request constructor

addPermission

Arguments

:: Text
apTopicArn

-> Text
apLabel

-> AddPermission

AddPermission constructor.

The fields accessible through corresponding lenses are:

apAWSAccountId :: [Text]

apActionName :: [Text]

apLabel :: Text

apTopicArn :: Text

Request lenses

apAWSAccountId :: Lens' AddPermission [Text]

The AWS account IDs of the users (principals) who will be given access to the specified actions. The users must have AWS accounts, but do not need to be signed up for this service.

apActionName :: Lens' AddPermission [Text]

The action you want to allow for the specified principal(s).

Valid values: any Amazon SNS action name.

apLabel :: Lens' AddPermission Text

A unique identifier for the new policy statement.

apTopicArn :: Lens' AddPermission Text

The ARN of the topic whose access control policy you wish to modify.

Response

data AddPermissionResponse

Response constructor

addPermissionResponse :: AddPermissionResponse

AddPermissionResponse constructor.