Network.AWS.SNS.CheckIfPhoneNumberIsOptedOut

Description

Accepts a phone number and indicates whether the phone holder has opted out of receiving SMS messages from your account. You cannot send SMS messages to a number that is opted out.

To resume sending messages, you can opt in the number by using the OptInPhoneNumber action.

Creating a Request

checkIfPhoneNumberIsOptedOut

Arguments

:: Text
cipniooPhoneNumber

-> CheckIfPhoneNumberIsOptedOut

Creates a value of CheckIfPhoneNumberIsOptedOut with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cipniooPhoneNumber - The phone number for which you want to check the opt out status.

data CheckIfPhoneNumberIsOptedOut

The input for the CheckIfPhoneNumberIsOptedOut action.

See: checkIfPhoneNumberIsOptedOut smart constructor.

Request Lenses

cipniooPhoneNumber :: Lens' CheckIfPhoneNumberIsOptedOut Text

The phone number for which you want to check the opt out status.

Destructuring the Response

checkIfPhoneNumberIsOptedOutResponse

Arguments

:: Int
cipnioorsResponseStatus

-> CheckIfPhoneNumberIsOptedOutResponse

Creates a value of CheckIfPhoneNumberIsOptedOutResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cipnioorsIsOptedOut - Indicates whether the phone number is opted out: * true – The phone number is opted out, meaning you cannot publish SMS messages to it. * false – The phone number is opted in, meaning you can publish SMS messages to it.

cipnioorsResponseStatus - -- | The response status code.

data CheckIfPhoneNumberIsOptedOutResponse

The response from the CheckIfPhoneNumberIsOptedOut action.

See: checkIfPhoneNumberIsOptedOutResponse smart constructor.

Response Lenses

cipnioorsIsOptedOut :: Lens' CheckIfPhoneNumberIsOptedOutResponse (Maybe Bool)

Indicates whether the phone number is opted out: * true – The phone number is opted out, meaning you cannot publish SMS messages to it. * false – The phone number is opted in, meaning you can publish SMS messages to it.

cipnioorsResponseStatus :: Lens' CheckIfPhoneNumberIsOptedOutResponse Int

- | The response status code.