Network.AWS.SNS.ConfirmSubscription

Description

Verifies an endpoint owner's intent to receive messages by validating the token sent to the endpoint by an earlier Subscribe action. If the token is valid, the action creates a new subscription and returns its Amazon Resource Name (ARN). This call requires an AWS signature only when the AuthenticateOnUnsubscribe flag is set to "true".

http://docs.aws.amazon.com/sns/latest/api/API_ConfirmSubscription.html

Request

data ConfirmSubscription

Request constructor

confirmSubscription

Arguments

:: Text
csTopicArn

-> Text
csToken

-> ConfirmSubscription

ConfirmSubscription constructor.

The fields accessible through corresponding lenses are:

csAuthenticateOnUnsubscribe :: Maybe Text

csToken :: Text

csTopicArn :: Text

Request lenses

csAuthenticateOnUnsubscribe :: Lens' ConfirmSubscription (Maybe Text)

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an AWS signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires AWS authentication.

csToken :: Lens' ConfirmSubscription Text

Short-lived token sent to an endpoint during the Subscribe action.

csTopicArn :: Lens' ConfirmSubscription Text

The ARN of the topic for which you wish to confirm a subscription.

Response

data ConfirmSubscriptionResponse

Response constructor

confirmSubscriptionResponse :: ConfirmSubscriptionResponse

ConfirmSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

csrSubscriptionArn :: Maybe Text

Response lenses

csrSubscriptionArn :: Lens' ConfirmSubscriptionResponse (Maybe Text)

The ARN of the created subscription.