Network.AWS.SNS.CreatePlatformApplication

Description

Creates a platform application object for one of the supported push notification services, such as APNS and GCM, to which devices and mobile apps may register. You must specify PlatformPrincipal and PlatformCredential attributes when using the CreatePlatformApplication action. The PlatformPrincipal is received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is "SSL certificate". For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". The PlatformCredential is also received from the notification service. For APNS/APNS_SANDBOX, PlatformCredential is "private key". For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". The PlatformApplicationArn that is returned when using CreatePlatformApplication is then used as an attribute for the CreatePlatformEndpoint action. For more information, see Using Amazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_CreatePlatformApplication.html

Request

data CreatePlatformApplication

Request constructor

createPlatformApplication

Arguments

:: Text
cpaName

-> Text
cpaPlatform

-> CreatePlatformApplication

CreatePlatformApplication constructor.

The fields accessible through corresponding lenses are:

cpaAttributes :: HashMap Text Text

cpaName :: Text

cpaPlatform :: Text

Request lenses

cpaAttributes :: Lens' CreatePlatformApplication (HashMap Text Text)

For a list of attributes, see SetPlatformApplicationAttributes

cpaName :: Lens' CreatePlatformApplication Text

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

cpaPlatform :: Lens' CreatePlatformApplication Text

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Google Cloud Messaging).

Response

data CreatePlatformApplicationResponse

Response constructor

createPlatformApplicationResponse :: CreatePlatformApplicationResponse

CreatePlatformApplicationResponse constructor.

The fields accessible through corresponding lenses are:

cparPlatformApplicationArn :: Maybe Text

Response lenses

cparPlatformApplicationArn :: Lens' CreatePlatformApplicationResponse (Maybe Text)

PlatformApplicationArn is returned.