Network.AWS.SNS.DeleteEndpoint

Description

Deletes the endpoint from Amazon SNS. This action is idempotent. For more information, see Using Amazon SNS Mobile Push Notifications. POST http:/sns.us-west-2.amazonaws.com HTTP/1.1 ... Action=DeleteEndpoint &amp;SignatureMethod=HmacSHA256 &amp;AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &amp;EndpointArn=arn%3Aaws%3Asns%3Aus-west-2%3A123456789012%3Aendpoint%2FGCM%2Fgcmpushapp%2F5e3e9847-3183-3f18-a7e8-671c3a57d4b3 &amp;SignatureVersion=2 &amp;Version=2010-03-31 &amp;Signature=LIc6GI3JbNhmHBEDmSxzZp648XPe5CMeFny%2BTQFtomQ%3D &amp;Timestamp=2013-07-01T23%3A00%3A12.456Z HTTP/1.1 200 OK ... &lt;DeleteEndpointResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;c1d2b191-353c-5a5f-8969-fbdd3900afa8&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;DeleteEndpointResponse&gt;.

Request

data DeleteEndpoint

Input for DeleteEndpoint action.

Request constructor

mkDeleteEndpoint

Arguments

:: Text
deEndpointArn

-> DeleteEndpoint

Smart constructor for the minimum required parameters to construct a valid DeleteEndpoint request.

The fields accessible through corresponding lenses are:

EndpointArn :: Text

Request lenses

deEndpointArn :: Lens' DeleteEndpoint Text

EndpointArn of endpoint to delete.

Response

data DeleteEndpointResponse

Response constructor

mkDeleteEndpointResponse :: DeleteEndpointResponse

Smart constructor for the minimum required parameters to construct a valid DeleteEndpointResponse response.

This constructor is provided for convenience and testing purposes.