Network.AWS.SNS.GetEndpointAttributes

Description

Retrieves the endpoint attributes for a device on one of the supported push notification services, such as GCM and APNS. For more information, see UsingAmazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_GetEndpointAttributes.html

Request

data GetEndpointAttributes

Request constructor

getEndpointAttributes

Arguments

:: Text
geaEndpointArn

-> GetEndpointAttributes

GetEndpointAttributes constructor.

The fields accessible through corresponding lenses are:

geaEndpointArn :: Text

Request lenses

geaEndpointArn :: Lens' GetEndpointAttributes Text

EndpointArn for GetEndpointAttributes input.

Response

data GetEndpointAttributesResponse

Response constructor

getEndpointAttributesResponse :: GetEndpointAttributesResponse

GetEndpointAttributesResponse constructor.

The fields accessible through corresponding lenses are:

gearAttributes :: HashMap Text Text

Response lenses

gearAttributes :: Lens' GetEndpointAttributesResponse (HashMap Text Text)

Attributes include the following:

CustomUserData -- arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB. Enabled -- flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token. Token -- device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.