Network.AWS.SNS.GetSubscriptionAttributes

Description

Returns all of the properties of a subscription.

http://docs.aws.amazon.com/sns/latest/api/API_GetSubscriptionAttributes.html

Request

data GetSubscriptionAttributes

Request constructor

getSubscriptionAttributes

Arguments

:: Text
gsaSubscriptionArn

-> GetSubscriptionAttributes

GetSubscriptionAttributes constructor.

The fields accessible through corresponding lenses are:

gsaSubscriptionArn :: Text

Request lenses

gsaSubscriptionArn :: Lens' GetSubscriptionAttributes Text

The ARN of the subscription whose properties you want to get.

Response

data GetSubscriptionAttributesResponse

Response constructor

getSubscriptionAttributesResponse :: GetSubscriptionAttributesResponse

GetSubscriptionAttributesResponse constructor.

The fields accessible through corresponding lenses are:

gsarAttributes :: HashMap Text Text

Response lenses

gsarAttributes :: Lens' GetSubscriptionAttributesResponse (HashMap Text Text)

A map of the subscription's attributes. Attributes in this map include the following: SubscriptionArn -- the subscription's ARN TopicArn -- the topic ARN that the subscription is associated with Owner -- the AWS account ID of the subscription's owner ConfirmationWasAuthenticated -- true if the subscription confirmation request was authenticated DeliveryPolicy -- the JSON serialization of the subscription's delivery policy EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account the topic delivery policy and account system defaults.