Network.AWS.SNS.GetTopicAttributes

Description

Returns all of the properties of a topic. Topic properties returned might differ based on the authorization of the user.

http://docs.aws.amazon.com/sns/latest/api/API_GetTopicAttributes.html

Request

data GetTopicAttributes

Request constructor

getTopicAttributes

Arguments

:: Text
gtaTopicArn

-> GetTopicAttributes

GetTopicAttributes constructor.

The fields accessible through corresponding lenses are:

gtaTopicArn :: Text

Request lenses

gtaTopicArn :: Lens' GetTopicAttributes Text

The ARN of the topic whose properties you want to get.

Response

data GetTopicAttributesResponse

Response constructor

getTopicAttributesResponse :: GetTopicAttributesResponse

GetTopicAttributesResponse constructor.

The fields accessible through corresponding lenses are:

gtarAttributes :: HashMap Text Text

Response lenses

gtarAttributes :: Lens' GetTopicAttributesResponse (HashMap Text Text)

A map of the topic's attributes. Attributes in this map include the following:

TopicArn -- the topic's ARN Owner -- the AWS account ID of the topic's owner Policy -- the JSON serialization of the topic's access control policy DisplayName -- the human-readable name used in the From field for notifications to email and email-json endpoints SubscriptionsPending -- the number of subscriptions pending confirmation on this topic SubscriptionsConfirmed -- the number of confirmed subscriptions on this topic SubscriptionsDeleted -- the number of deleted subscriptions on this topic DeliveryPolicy -- the JSON serialization of the topic's delivery policy EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account system defaults