Network.AWS.SNS.ListEndpointsByPlatformApplication

Description

Lists the endpoints and endpoint attributes for devices in a supported push notification service, such as GCM and APNS. The results for ListEndpointsByPlatformApplication are paginated and return a limited list of endpoints, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListEndpointsByPlatformApplication again using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_ListEndpointsByPlatformApplication.html

Request

data ListEndpointsByPlatformApplication

Request constructor

listEndpointsByPlatformApplication

Arguments

:: Text
lebpaPlatformApplicationArn

-> ListEndpointsByPlatformApplication

ListEndpointsByPlatformApplication constructor.

The fields accessible through corresponding lenses are:

lebpaNextToken :: Maybe Text

lebpaPlatformApplicationArn :: Text

Request lenses

lebpaNextToken :: Lens' ListEndpointsByPlatformApplication (Maybe Text)

NextToken string is used when calling ListEndpointsByPlatformApplication action to retrieve additional records that are available after the first page results.

lebpaPlatformApplicationArn :: Lens' ListEndpointsByPlatformApplication Text

PlatformApplicationArn for ListEndpointsByPlatformApplicationInput action.

Response

data ListEndpointsByPlatformApplicationResponse

Response constructor

listEndpointsByPlatformApplicationResponse :: ListEndpointsByPlatformApplicationResponse

ListEndpointsByPlatformApplicationResponse constructor.

The fields accessible through corresponding lenses are:

lebparEndpoints :: [Endpoint]

lebparNextToken :: Maybe Text

Response lenses

lebparEndpoints :: Lens' ListEndpointsByPlatformApplicationResponse [Endpoint]

Endpoints returned for ListEndpointsByPlatformApplication action.

lebparNextToken :: Lens' ListEndpointsByPlatformApplicationResponse (Maybe Text)

NextToken string is returned when calling ListEndpointsByPlatformApplication action if additional records are available after the first page results.