Network.AWS.SNS.ListPlatformApplications

Description

Lists the platform application objects for the supported push notification services, such as APNS and GCM. The results for ListPlatformApplications are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListPlatformApplications using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_ListPlatformApplications.html

Request

data ListPlatformApplications

Request constructor

listPlatformApplications :: ListPlatformApplications

ListPlatformApplications constructor.

The fields accessible through corresponding lenses are:

lpaNextToken :: Maybe Text

Request lenses

lpaNextToken :: Lens' ListPlatformApplications (Maybe Text)

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

Response

data ListPlatformApplicationsResponse

Response constructor

listPlatformApplicationsResponse :: ListPlatformApplicationsResponse

ListPlatformApplicationsResponse constructor.

The fields accessible through corresponding lenses are:

lparNextToken :: Maybe Text

lparPlatformApplications :: [PlatformApplication]

Response lenses

lparNextToken :: Lens' ListPlatformApplicationsResponse (Maybe Text)

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

lparPlatformApplications :: Lens' ListPlatformApplicationsResponse [PlatformApplication]

Platform applications returned when calling ListPlatformApplications action.