Network.AWS.SNS.ListSubscriptions

Description

Returns a list of the requester's subscriptions. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptions call to get further results.

http://docs.aws.amazon.com/sns/latest/api/API_ListSubscriptions.html

Request

data ListSubscriptions

Request constructor

listSubscriptions :: ListSubscriptions

ListSubscriptions constructor.

The fields accessible through corresponding lenses are:

lsNextToken :: Maybe Text

Request lenses

lsNextToken :: Lens' ListSubscriptions (Maybe Text)

Token returned by the previous ListSubscriptions request.

Response

data ListSubscriptionsResponse

Response constructor

listSubscriptionsResponse :: ListSubscriptionsResponse

ListSubscriptionsResponse constructor.

The fields accessible through corresponding lenses are:

lsrNextToken :: Maybe Text

lsrSubscriptions :: [Subscription]

Response lenses

lsrNextToken :: Lens' ListSubscriptionsResponse (Maybe Text)

Token to pass along to the next ListSubscriptions request. This element is returned if there are more subscriptions to retrieve.

lsrSubscriptions :: Lens' ListSubscriptionsResponse [Subscription]

A list of subscriptions.