Network.AWS.SNS.ListSubscriptionsByTopic

Description

Returns a list of the subscriptions to a specific topic. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptionsByTopic call to get further results.

http://docs.aws.amazon.com/sns/latest/api/API_ListSubscriptionsByTopic.html

Request

data ListSubscriptionsByTopic

Request constructor

listSubscriptionsByTopic

Arguments

:: Text
lsbtTopicArn

-> ListSubscriptionsByTopic

ListSubscriptionsByTopic constructor.

The fields accessible through corresponding lenses are:

lsbtNextToken :: Maybe Text

lsbtTopicArn :: Text

Request lenses

lsbtNextToken :: Lens' ListSubscriptionsByTopic (Maybe Text)

Token returned by the previous ListSubscriptionsByTopic request.

lsbtTopicArn :: Lens' ListSubscriptionsByTopic Text

The ARN of the topic for which you wish to find subscriptions.

Response

data ListSubscriptionsByTopicResponse

Response constructor

listSubscriptionsByTopicResponse :: ListSubscriptionsByTopicResponse

ListSubscriptionsByTopicResponse constructor.

The fields accessible through corresponding lenses are:

lsbtrNextToken :: Maybe Text

lsbtrSubscriptions :: [Subscription]

Response lenses

lsbtrNextToken :: Lens' ListSubscriptionsByTopicResponse (Maybe Text)

Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.

lsbtrSubscriptions :: Lens' ListSubscriptionsByTopicResponse [Subscription]

A list of subscriptions.