Network.AWS.SNS.ListTopics

Description

Returns a list of the requester's topics. Each call returns a limited list of topics, up to 100. If there are more topics, a NextToken is also returned. Use the NextToken parameter in a new ListTopics call to get further results.

http://docs.aws.amazon.com/sns/latest/api/API_ListTopics.html

Request

data ListTopics

Request constructor

listTopics :: ListTopics

ListTopics constructor.

The fields accessible through corresponding lenses are:

ltNextToken :: Maybe Text

Request lenses

ltNextToken :: Lens' ListTopics (Maybe Text)

Token returned by the previous ListTopics request.

Response

data ListTopicsResponse

Response constructor

listTopicsResponse :: ListTopicsResponse

ListTopicsResponse constructor.

The fields accessible through corresponding lenses are:

ltrNextToken :: Maybe Text

ltrTopics :: [Topic]

Response lenses

ltrNextToken :: Lens' ListTopicsResponse (Maybe Text)

Token to pass along to the next ListTopics request. This element is returned if there are additional topics to retrieve.

ltrTopics :: Lens' ListTopicsResponse [Topic]

A list of topic ARNs.