Network.AWS.SNS.SetEndpointAttributes

Description

Sets the attributes for an endpoint for a device on one of the supported push notification services, such as GCM and APNS. For more information, see UsingAmazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_SetEndpointAttributes.html

Request

data SetEndpointAttributes

Request constructor

setEndpointAttributes

Arguments

:: Text
seaEndpointArn

-> SetEndpointAttributes

SetEndpointAttributes constructor.

The fields accessible through corresponding lenses are:

seaAttributes :: HashMap Text Text

seaEndpointArn :: Text

Request lenses

seaAttributes :: Lens' SetEndpointAttributes (HashMap Text Text)

A map of the endpoint attributes. Attributes in this map include the following:

CustomUserData -- arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB. Enabled -- flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token. Token -- device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

seaEndpointArn :: Lens' SetEndpointAttributes Text

EndpointArn used for SetEndpointAttributes action.

Response

data SetEndpointAttributesResponse

Response constructor

setEndpointAttributesResponse :: SetEndpointAttributesResponse

SetEndpointAttributesResponse constructor.