Network.AWS.SNS.SetPlatformApplicationAttributes

Description

Sets the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see UsingAmazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_SetPlatformApplicationAttributes.html

Request

data SetPlatformApplicationAttributes

Request constructor

setPlatformApplicationAttributes

Arguments

:: Text
spaaPlatformApplicationArn

-> SetPlatformApplicationAttributes

SetPlatformApplicationAttributes constructor.

The fields accessible through corresponding lenses are:

spaaAttributes :: HashMap Text Text

spaaPlatformApplicationArn :: Text

Request lenses

spaaAttributes :: Lens' SetPlatformApplicationAttributes (HashMap Text Text)

A map of the platform application attributes. Attributes in this map include the following:

PlatformCredential -- The credential received from the notification service. For APNS/APNS_SANDBOX, PlatformCredential is "private key". For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". PlatformPrincipal -- The principal received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is "SSL certificate". For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

spaaPlatformApplicationArn :: Lens' SetPlatformApplicationAttributes Text

PlatformApplicationArn for SetPlatformApplicationAttributes action.

Response

data SetPlatformApplicationAttributesResponse

Response constructor

setPlatformApplicationAttributesResponse :: SetPlatformApplicationAttributesResponse

SetPlatformApplicationAttributesResponse constructor.