Network.AWS.SNS.SetSubscriptionAttributes

Description

Allows a subscription owner to set an attribute of the topic to a new value. The following example sets the delivery policy to 5 total retries http:/sns.us-east-1.amazonaws.com ?AttributeValue={"healthyRetryPolicy":{"numRetries":5}} &amp;SubscriptionArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Topic%3A80289ba6-0fd4-4079-afb4-ce8c8260f0ca &amp;AttributeName=DeliveryPolicy &amp;Action=SetSubscriptionAttributes &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key Id) &amp;Signature=mQA3nJI%2BcmAIY7r8HCArGElSqPX5JG4UGzF4yo0RygE%3D The JSON format for the DeliveryPolicy AttributeValue (linebreaks added for readability): { "healthyRetryPolicy": { "minDelayTarget": &lt;int&gt;, "maxDelayTarget": &lt;int&gt;, "numRetries": &lt;int&gt;, "numMaxDelayRetries": &lt;int&gt;, "backoffFunction": "&lt;linear|arithmetic|geometric|exponential&gt;" }, "throttlePolicy": { "maxReceivesPerSecond": &lt;int&gt; } } &lt;SetSubscriptionAttributesResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;a8763b99-33a7-11df-a9b7-05d48da6f042&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;SetSubscriptionAttributesResponse&gt;.

Request

data SetSubscriptionAttributes

Input for SetSubscriptionAttributes action.

Request constructor

setSubscriptionAttributes

Arguments

:: Text
ssaSubscriptionArn

-> Text
ssaAttributeName

-> SetSubscriptionAttributes

Smart constructor for the minimum required parameters to construct a valid SetSubscriptionAttributes request.

The fields accessible through corresponding lenses are:

SubscriptionArn :: Text

AttributeName :: Text

AttributeValue :: Maybe Text

Request lenses

ssaSubscriptionArn :: Lens' SetSubscriptionAttributes Text

The ARN of the subscription to modify.

ssaAttributeName :: Lens' SetSubscriptionAttributes Text

The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable. Valid values: DeliveryPolicy | RawMessageDelivery.

ssaAttributeValue :: Lens' SetSubscriptionAttributes (Maybe Text)

The new value for the attribute in JSON format.

Response

data SetSubscriptionAttributesResponse

Response constructor

setSubscriptionAttributesResponse :: SetSubscriptionAttributesResponse

Smart constructor for the minimum required parameters to construct a valid SetSubscriptionAttributesResponse response.

This constructor is provided for convenience and testing purposes.