Network.AWS.SNS.SetTopicAttributes

Description

Allows a topic owner to set an attribute of the topic to a new value.

http://docs.aws.amazon.com/sns/latest/api/API_SetTopicAttributes.html

Request

data SetTopicAttributes

Request constructor

setTopicAttributes

Arguments

:: Text
staTopicArn

-> Text
staAttributeName

-> SetTopicAttributes

SetTopicAttributes constructor.

The fields accessible through corresponding lenses are:

staAttributeName :: Text

staAttributeValue :: Maybe Text

staTopicArn :: Text

Request lenses

staAttributeName :: Lens' SetTopicAttributes Text

The name of the attribute you want to set. Only a subset of the topic's attributes are mutable.

Valid values: Policy | DisplayName | DeliveryPolicy

staAttributeValue :: Lens' SetTopicAttributes (Maybe Text)

The new value for the attribute.

staTopicArn :: Lens' SetTopicAttributes Text

The ARN of the topic to modify.

Response

data SetTopicAttributesResponse

Response constructor

setTopicAttributesResponse :: SetTopicAttributesResponse

SetTopicAttributesResponse constructor.