Network.AWS.SNS.Subscribe

Description

Prepares to subscribe an endpoint by sending the endpoint a confirmation message. To actually create a subscription, the endpoint owner must call the ConfirmSubscription action with the token from the confirmation message. Confirmation tokens are valid for three days.

http://docs.aws.amazon.com/sns/latest/api/API_Subscribe.html

Request

data Subscribe

Request constructor

subscribe

Arguments

:: Text
sTopicArn

-> Text
sProtocol

-> Subscribe

Subscribe constructor.

The fields accessible through corresponding lenses are:

sEndpoint :: Maybe Text

sProtocol :: Text

sTopicArn :: Text

Request lenses

sEndpoint :: Lens' Subscribe (Maybe Text)

The endpoint that you want to receive notifications. Endpoints vary by protocol:

For the http protocol, the endpoint is an URL beginning with "http://" For the https protocol, the endpoint is a URL beginning with "https://" For the email protocol, the endpoint is an email address For the 'email-json' protocol, the endpoint is an email address For the sms protocol, the endpoint is a phone number of an SMS-enabled device For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

sProtocol :: Lens' Subscribe Text

The protocol you want to use. Supported protocols include:

http -- delivery of JSON-encoded message via HTTP POST https -- delivery of JSON-encoded message via HTTPS POST email -- delivery of message via SMTP 'email-json' -- delivery of JSON-encoded message via SMTP sms -- delivery of message via SMS sqs -- delivery of JSON-encoded message to an Amazon SQS queue application -- delivery of JSON-encoded message to an EndpointArn for a mobile app and device.

sTopicArn :: Lens' Subscribe Text

The ARN of the topic you want to subscribe to.

Response

data SubscribeResponse

Response constructor

subscribeResponse :: SubscribeResponse

SubscribeResponse constructor.

The fields accessible through corresponding lenses are:

srSubscriptionArn :: Maybe Text

Response lenses

srSubscriptionArn :: Lens' SubscribeResponse (Maybe Text)

The ARN of the subscription, if the service was able to create a subscription immediately (without requiring endpoint owner confirmation).