Network.AWS.SNS.Types

Description

Amazon Simple Notification Service (Amazon SNS) is a fast, flexible, fully managed push messaging service. Amazon SNS makes it simple and cost-effective to push notifications to Apple, Google, Fire OS, and Windows devices, as well as Android devices in China with Baidu Cloud Push. You can also use SNS to push notifications to internet connected smart devices, as well as other distributed services. Besides pushing cloud notifications directly to mobile devices, Amazon SNS can also deliver notifications by SMS text message or email, to Amazon Simple Queue Service (SQS) queues, or to any HTTP endpoint. To prevent messages from being lost, all messages published to Amazon SNS are stored redundantly across multiple availability zones.

Service

data SNS

Supported version (2010-03-31) of the Amazon Simple Notification Service service.

Errors

data SNSError

A sum type representing possible errors returned by the SNS service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

AuthorizationErrorException
Indicates that the user has been denied access to the requested resource.

Fields

_aeeMessage :: Maybe Text

EndpointDisabledException
Exception error indicating endpoint disabled.

Fields

_edeMessage :: Maybe Text

InternalErrorException
Indicates an internal service error.

Fields

_ieeMessage :: Maybe Text

InvalidParameterException
Indicates that a request parameter does not comply with the associated constraints.

Fields

_ipeMessage :: Maybe Text

InvalidParameterValueException
Indicates that a request parameter does not comply with the associated constraints.

Fields

_ipveMessage :: Maybe Text

NotFoundException
Indicates that the requested resource does not exist.

Fields

_nfeMessage :: Maybe Text

PlatformApplicationDisabledException
Exception error indicating platform application disabled.

Fields

_padeMessage :: Maybe Text

SNSClient HttpException

SNSSerializer String

SNSService String

SubscriptionLimitExceededException
Indicates that the customer already owns the maximum allowed number of subscriptions.

Fields

_sleeMessage :: Maybe Text

TopicLimitExceededException
Indicates that the customer already owns the maximum allowed number of topics.

Fields

_tleeMessage :: Maybe Text

_AuthorizationErrorException :: Prism' SNSError (Maybe Text)

Indicates that the user has been denied access to the requested resource.

See: AuthorizationErrorException

_EndpointDisabledException :: Prism' SNSError (Maybe Text)

Exception error indicating endpoint disabled.

See: EndpointDisabledException

_InternalErrorException :: Prism' SNSError (Maybe Text)

Indicates an internal service error.

See: InternalErrorException

_InvalidParameterException :: Prism' SNSError (Maybe Text)

Indicates that a request parameter does not comply with the associated constraints.

See: InvalidParameterException

_InvalidParameterValueException :: Prism' SNSError (Maybe Text)

Indicates that a request parameter does not comply with the associated constraints.

See: InvalidParameterValueException

_NotFoundException :: Prism' SNSError (Maybe Text)

Indicates that the requested resource does not exist.

See: NotFoundException

_PlatformApplicationDisabledException :: Prism' SNSError (Maybe Text)

Exception error indicating platform application disabled.

See: PlatformApplicationDisabledException

_SNSClient :: Prism' SNSError HttpException

See: SNSClient

_SNSSerializer :: Prism' SNSError String

See: SNSSerializer

_SNSService :: Prism' SNSError String

See: SNSService

_SubscriptionLimitExceededException :: Prism' SNSError (Maybe Text)

Indicates that the customer already owns the maximum allowed number of subscriptions.

See: SubscriptionLimitExceededException

_TopicLimitExceededException :: Prism' SNSError (Maybe Text)

Indicates that the customer already owns the maximum allowed number of topics.

See: TopicLimitExceededException

XML

xmlOptions :: Tagged a XMLOptions

Topic

data Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes.

topic :: Topic

Smart constructor for the minimum required fields to construct a valid Topic data type.

Topic is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TopicArn :: Maybe Text

tTopicArn :: Lens' Topic (Maybe Text)

The topic's ARN.

Endpoint'

data Endpoint'

Endpoint for mobile app and device.

endpoint' :: Endpoint'

Smart constructor for the minimum required fields to construct a valid Endpoint' data type.

Endpoint' is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EndpointArn :: Maybe Text

Attributes :: Map Text Text

eEndpointArn :: Lens' Endpoint' (Maybe Text)

EndpointArn for mobile app and device.

eAttributes :: Lens' Endpoint' (Map Text Text)

Attributes for endpoint.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Using Amazon SNS Message Attributes.

messageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

Smart constructor for the minimum required fields to construct a valid MessageAttributeValue data type to populate a request.

The fields accessible through corresponding lenses are:

DataType :: Text

StringValue :: Maybe Text

BinaryValue :: Maybe ByteString

mavDataType :: Lens' MessageAttributeValue Text

Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types.

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

PlatformApplication

data PlatformApplication

Platform application object.

platformApplication :: PlatformApplication

Smart constructor for the minimum required fields to construct a valid PlatformApplication data type.

PlatformApplication is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PlatformApplicationArn :: Maybe Text

Attributes :: Map Text Text

paPlatformApplicationArn :: Lens' PlatformApplication (Maybe Text)

PlatformApplicationArn for platform application object.

paAttributes :: Lens' PlatformApplication (Map Text Text)

Attributes for platform application object.

Subscription

data Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

subscription :: Subscription

Smart constructor for the minimum required fields to construct a valid Subscription data type.

Subscription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SubscriptionArn :: Maybe Text

Owner :: Maybe Text

Protocol :: Maybe Text

Endpoint :: Maybe Text

TopicArn :: Maybe Text

sSubscriptionArn :: Lens' Subscription (Maybe Text)

The subscription's ARN.

sOwner :: Lens' Subscription (Maybe Text)

The subscription's owner.

sProtocol :: Lens' Subscription (Maybe Text)

The subscription's protocol.

sEndpoint :: Lens' Subscription (Maybe Text)

The subscription's endpoint (format depends on the protocol).

sTopicArn :: Lens' Subscription (Maybe Text)

The ARN of the subscription's topic.