Network.AWS.SNS.Unsubscribe

Description

Deletes a subscription. If the subscription requires authentication for deletion, only the owner of the subscription or the topic's owner can unsubscribe, and an AWS signature is required. If the Unsubscribe call does not require authentication and the requester is not the subscription owner, a final cancellation message is delivered to the endpoint, so that the endpoint owner can easily resubscribe to the topic if the Unsubscribe request was unintended.

http://docs.aws.amazon.com/sns/latest/api/API_Unsubscribe.html

Request

data Unsubscribe

Request constructor

unsubscribe

Arguments

:: Text
uSubscriptionArn

-> Unsubscribe

Unsubscribe constructor.

The fields accessible through corresponding lenses are:

uSubscriptionArn :: Text

Request lenses

uSubscriptionArn :: Lens' Unsubscribe Text

The ARN of the subscription to be deleted.

Response

data UnsubscribeResponse

Response constructor

unsubscribeResponse :: UnsubscribeResponse

UnsubscribeResponse constructor.