Network.AWS.SNS.V2010_03_31

Description

Amazon Simple Notification Service (Amazon SNS) is a fast, flexible, fully managed push messaging service. Amazon SNS makes it simple and cost-effective to push notifications to Apple, Google, Fire OS, and Windows devices, as well as Android devices in China with Baidu Cloud Push. You can also use SNS to push notifications to internet connected smart devices, as well as other distributed services. Besides pushing cloud notifications directly to mobile devices, Amazon SNS can also deliver notifications by SMS text message or email, to Amazon Simple Queue Service (SQS) queues, or to any HTTP endpoint. To prevent messages from being lost, all messages published to Amazon SNS are stored redundantly across multiple availability zones.

Documentation

data AddPermission

Constructors

AddPermission

Fields

_apiActionName :: [Text]

The action you want to allow for the specified principal(s). Valid values: any Amazon SNS action name.

_apiAWSAccountId :: [Text]

The AWS account IDs of the users (principals) who will be given access to the specified actions. The users must have AWS accounts, but do not need to be signed up for this service.

_apiLabel :: Text

A unique identifier for the new policy statement.

_apiTopicArn :: Text

The ARN of the topic whose access control policy you wish to modify.

apiTopicArn :: Lens' AddPermission Text

apiLabel :: Lens' AddPermission Text

apiActionName :: Lens' AddPermission [Text]

apiAWSAccountId :: Lens' AddPermission [Text]

data AddPermissionResponse

Constructors

AddPermissionResponse

confirmSubscription

Arguments

:: Text
_csiToken

-> Text
_csiTopicArn

-> ConfirmSubscription

Minimum specification for a ConfirmSubscription request.

data ConfirmSubscription

Constructors

ConfirmSubscription

Fields

_csiToken :: Text

Short-lived token sent to an endpoint during the Subscribe action.

_csiTopicArn :: Text

The ARN of the topic for which you wish to confirm a subscription.

_csiAuthenticateOnUnsubscribe :: Maybe Text

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an AWS signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires AWS authentication.

csiTopicArn :: Lens' ConfirmSubscription Text

csiToken :: Lens' ConfirmSubscription Text

csiAuthenticateOnUnsubscribe :: Lens' ConfirmSubscription (Maybe Text)

data ConfirmSubscriptionResponse

Constructors

ConfirmSubscriptionResponse

Fields

_csrSubscriptionArn :: Maybe Text

The ARN of the created subscription.

csrSubscriptionArn :: Iso' ConfirmSubscriptionResponse (Maybe Text)

data CreatePlatformApplication

Constructors

CreatePlatformApplication

Fields

_cpaiAttributes :: Map Text Text

For a list of attributes, see SetPlatformApplicationAttributes.

_cpaiPlatform :: Text

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Google Cloud Messaging).

_cpaiName :: Text

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

cpaiPlatform :: Lens' CreatePlatformApplication Text

cpaiName :: Lens' CreatePlatformApplication Text

cpaiAttributes :: Lens' CreatePlatformApplication (Map Text Text)

data CreatePlatformApplicationResponse

Constructors

CreatePlatformApplicationResponse

Fields

_cparPlatformApplicationArn :: Maybe Text

PlatformApplicationArn is returned.

cparPlatformApplicationArn :: Iso' CreatePlatformApplicationResponse (Maybe Text)

createPlatformEndpoint

Arguments

:: Text
_cpeiPlatformApplicationArn

-> Text
_cpeiToken

-> CreatePlatformEndpoint

Minimum specification for a CreatePlatformEndpoint request.

data CreatePlatformEndpoint

Constructors

CreatePlatformEndpoint

Fields

_cpeiPlatformApplicationArn :: Text

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

_cpeiToken :: Text

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM or ADM, the device token equivalent is called the registration ID.

_cpeiAttributes :: Map Text Text

For a list of attributes, see SetEndpointAttributes.

_cpeiCustomUserData :: Maybe Text

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

cpeiToken :: Lens' CreatePlatformEndpoint Text

cpeiPlatformApplicationArn :: Lens' CreatePlatformEndpoint Text

cpeiCustomUserData :: Lens' CreatePlatformEndpoint (Maybe Text)

cpeiAttributes :: Lens' CreatePlatformEndpoint (Map Text Text)

data CreatePlatformEndpointResponse

Constructors

CreatePlatformEndpointResponse

Fields

_cerEndpointArn :: Maybe Text

EndpointArn returned from CreateEndpoint action.

cerEndpointArn :: Iso' CreatePlatformEndpointResponse (Maybe Text)

data CreateTopic

Constructors

CreateTopic

Fields

_ctiName :: Text

The name of the topic you want to create. Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

ctiName :: Iso' CreateTopic Text

data CreateTopicResponse

Constructors

CreateTopicResponse

Fields

_ctrTopicArn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the created topic.

ctrTopicArn :: Iso' CreateTopicResponse (Maybe Text)

data DeleteEndpoint

Constructors

DeleteEndpoint

Fields

_deiEndpointArn :: Text

EndpointArn of endpoint to delete.

deiEndpointArn :: Iso' DeleteEndpoint Text

data DeleteEndpointResponse

Constructors

DeleteEndpointResponse

data DeletePlatformApplication

Constructors

DeletePlatformApplication

Fields

_dpaiPlatformApplicationArn :: Text

PlatformApplicationArn of platform application object to delete.

dpaiPlatformApplicationArn :: Iso' DeletePlatformApplication Text

data DeletePlatformApplicationResponse

Constructors

DeletePlatformApplicationResponse

data DeleteTopic

Constructors

DeleteTopic

Fields

_dtiTopicArn :: Text

The ARN of the topic you want to delete. http:/sns.us-east-1.amazonaws.com ?TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Topic &amp;Action=DeleteTopic &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=DjHBa%2BbYCKQAzctOPnLP7MbHnrHT3%2FK3kFEZjwcf9%2FU%3D &lt;DeleteTopicResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;fba800b9-3765-11df-8cf3-c58c53254dfb&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;DeleteTopicResponse&gt;.

dtiTopicArn :: Iso' DeleteTopic Text

data DeleteTopicResponse

Constructors

DeleteTopicResponse

data GetEndpointAttributes

Constructors

GetEndpointAttributes

Fields

_geaiEndpointArn :: Text

EndpointArn for GetEndpointAttributes input.

geaiEndpointArn :: Iso' GetEndpointAttributes Text

data GetEndpointAttributesResponse

Constructors

GetEndpointAttributesResponse

Fields

_gearAttributes :: Map Text Text

Attributes include the following: CustomUserData -- arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB. Enabled -- flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token. Token -- device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

gearAttributes :: Iso' GetEndpointAttributesResponse (Map Text Text)

data GetPlatformApplicationAttributes

Constructors

GetPlatformApplicationAttributes

Fields

_gpaaiPlatformApplicationArn :: Text

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

gpaaiPlatformApplicationArn :: Iso' GetPlatformApplicationAttributes Text

data GetPlatformApplicationAttributesResponse

Constructors

GetPlatformApplicationAttributesResponse

Fields

_gpaarAttributes :: Map Text Text

Attributes include the following: EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

gpaarAttributes :: Iso' GetPlatformApplicationAttributesResponse (Map Text Text)

data GetSubscriptionAttributes

Constructors

GetSubscriptionAttributes

Fields

_gsaiSubscriptionArn :: Text

The ARN of the subscription whose properties you want to get.

gsaiSubscriptionArn :: Iso' GetSubscriptionAttributes Text

data GetSubscriptionAttributesResponse

Constructors

GetSubscriptionAttributesResponse

Fields

_gsarAttributes :: Map Text Text

A map of the subscription's attributes. Attributes in this map include the following: SubscriptionArn -- the subscription's ARN TopicArn -- the topic ARN that the subscription is associated with Owner -- the AWS account ID of the subscription's owner ConfirmationWasAuthenticated -- true if the subscription confirmation request was authenticated DeliveryPolicy -- the JSON serialization of the subscription's delivery policy EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account the topic delivery policy and account system defaults.

gsarAttributes :: Iso' GetSubscriptionAttributesResponse (Map Text Text)

data GetTopicAttributes

Constructors

GetTopicAttributes

Fields

_gtaiTopicArn :: Text

The ARN of the topic whose properties you want to get.

gtaiTopicArn :: Iso' GetTopicAttributes Text

data GetTopicAttributesResponse

Constructors

GetTopicAttributesResponse

Fields

_gtarAttributes :: Map Text Text

A map of the topic's attributes. Attributes in this map include the following: TopicArn -- the topic's ARN Owner -- the AWS account ID of the topic's owner Policy -- the JSON serialization of the topic's access control policy DisplayName -- the human-readable name used in the From field for notifications to email and email-json endpoints SubscriptionsPending -- the number of subscriptions pending confirmation on this topic SubscriptionsConfirmed -- the number of confirmed subscriptions on this topic SubscriptionsDeleted -- the number of deleted subscriptions on this topic DeliveryPolicy -- the JSON serialization of the topic's delivery policy EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account system defaults.

gtarAttributes :: Iso' GetTopicAttributesResponse (Map Text Text)

listEndpointsByPlatformApplication

Arguments

:: Text
_lebpaiPlatformApplicationArn

-> ListEndpointsByPlatformApplication

Minimum specification for a ListEndpointsByPlatformApplication request.

data ListEndpointsByPlatformApplication

Constructors

ListEndpointsByPlatformApplication

Fields

_lebpaiPlatformApplicationArn :: Text

PlatformApplicationArn for ListEndpointsByPlatformApplicationInput action.

_lebpaiNextToken :: Maybe Text

NextToken string is used when calling ListEndpointsByPlatformApplication action to retrieve additional records that are available after the first page results.

lebpaiPlatformApplicationArn :: Lens' ListEndpointsByPlatformApplication Text

lebpaiNextToken :: Lens' ListEndpointsByPlatformApplication (Maybe Text)

data ListEndpointsByPlatformApplicationResponse

Constructors

ListEndpointsByPlatformApplicationResponse

Fields

_lebparEndpoints :: [Endpoint]

Endpoints returned for ListEndpointsByPlatformApplication action.

_lebparNextToken :: Maybe Text

NextToken string is returned when calling ListEndpointsByPlatformApplication action if additional records are available after the first page results.

lebparNextToken :: Lens' ListEndpointsByPlatformApplicationResponse (Maybe Text)

lebparEndpoints :: Lens' ListEndpointsByPlatformApplicationResponse [Endpoint]

listPlatformApplications :: ListPlatformApplications

Minimum specification for a ListPlatformApplications request.

data ListPlatformApplications

Constructors

ListPlatformApplications

Fields

_lpaiNextToken :: Maybe Text

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

lpaiNextToken :: Iso' ListPlatformApplications (Maybe Text)

data ListPlatformApplicationsResponse

Constructors

ListPlatformApplicationsResponse

Fields

_lparPlatformApplications :: [PlatformApplication]

Platform applications returned when calling ListPlatformApplications action.

_lparNextToken :: Maybe Text

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

lparPlatformApplications :: Lens' ListPlatformApplicationsResponse [PlatformApplication]

lparNextToken :: Lens' ListPlatformApplicationsResponse (Maybe Text)

listSubscriptions :: ListSubscriptions

Minimum specification for a ListSubscriptions request.

data ListSubscriptions

Constructors

ListSubscriptions

Fields

_lsiNextToken :: Maybe Text

Token returned by the previous ListSubscriptions request.

lsiNextToken :: Iso' ListSubscriptions (Maybe Text)

data ListSubscriptionsResponse

Constructors

ListSubscriptionsResponse

Fields

_lsrNextToken :: Maybe Text

Token to pass along to the next ListSubscriptions request. This element is returned if there are more subscriptions to retrieve.

_lsrSubscriptions :: [Subscription]

A list of subscriptions.

lsrSubscriptions :: Lens' ListSubscriptionsResponse [Subscription]

lsrNextToken :: Lens' ListSubscriptionsResponse (Maybe Text)

listSubscriptionsByTopic

Arguments

:: Text
_lsbtiTopicArn

-> ListSubscriptionsByTopic

Minimum specification for a ListSubscriptionsByTopic request.

data ListSubscriptionsByTopic

Constructors

ListSubscriptionsByTopic

Fields

_lsbtiTopicArn :: Text

The ARN of the topic for which you wish to find subscriptions.

_lsbtiNextToken :: Maybe Text

Token returned by the previous ListSubscriptionsByTopic request.

lsbtiTopicArn :: Lens' ListSubscriptionsByTopic Text

lsbtiNextToken :: Lens' ListSubscriptionsByTopic (Maybe Text)

data ListSubscriptionsByTopicResponse

Constructors

ListSubscriptionsByTopicResponse

Fields

_lsbtrNextToken :: Maybe Text

Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.

_lsbtrSubscriptions :: [Subscription]

A list of subscriptions.

lsbtrSubscriptions :: Lens' ListSubscriptionsByTopicResponse [Subscription]

lsbtrNextToken :: Lens' ListSubscriptionsByTopicResponse (Maybe Text)

listTopics :: ListTopics

Minimum specification for a ListTopics request.

data ListTopics

Constructors

ListTopics

Fields

_ltiNextToken :: Maybe Text

Token returned by the previous ListTopics request.

ltiNextToken :: Iso' ListTopics (Maybe Text)

data ListTopicsResponse

Constructors

ListTopicsResponse

Fields

_ltrNextToken :: Maybe Text

Token to pass along to the next ListTopics request. This element is returned if there are additional topics to retrieve.

_ltrTopics :: [Topic]

A list of topic ARNs.

ltrTopics :: Lens' ListTopicsResponse [Topic]

ltrNextToken :: Lens' ListTopicsResponse (Maybe Text)

publish

Arguments

:: Text
_piMessage

-> Publish

Minimum specification for a Publish request.

data Publish

Constructors

Publish

Fields

_piMessage :: Text

The message you want to send to the topic. If you want to send the same message to all transport protocols, include the text of the message as a String value. If you want to send different messages for each transport protocol, set the value of the MessageStructure parameter to json and use a JSON object for the Message parameter. See the Examples section for the format of the JSON object. Constraints: Messages must be UTF-8 encoded strings at most 256 KB in size (262144 bytes, not 262144 characters). JSON-specific constraints: Keys in the JSON object that correspond to supported transport protocols must have simple JSON string values. The values will be parsed (unescaped) before they are used in outgoing messages. Outbound notifications are JSON encoded (meaning that the characters will be reescaped for sending). Values have a minimum length of 0 (the empty string, "", is allowed). Values have a maximum length bounded by the overall message size (so, including multiple protocols may limit message sizes). Non-string values will cause the key to be ignored. Keys that do not correspond to supported transport protocols are ignored. Duplicate keys are not allowed. Failure to parse or validate any key or value in the message will cause the Publish call to return an error (no partial delivery).

_piMessageAttributes :: Map Text MessageAttributeValue

Message attributes for Publish action.

_piMessageStructure :: Maybe Text

Set MessageStructure to json if you want to send a different message for each protocol. For example, using one publish action, you can send a short message to your SMS subscribers and a longer message to your email subscribers. If you set MessageStructure to json, the value of the Message parameter must: be a syntactically valid JSON object; and contain at least a top-level JSON key of "default" with a value that is a string. You can define other top-level keys that define the message you want to send to a specific transport protocol (e.g., "http"). For information about sending different messages for each protocol using the AWS Management Console, go to Create Different Messages for Each Protocol in the Amazon Simple Notification Service Getting Started Guide. Valid value: json.

_piTargetArn :: Maybe Text

Either TopicArn or EndpointArn, but not both.

_piSubject :: Maybe Text

Optional parameter to be used as the Subject line when the message is delivered to email endpoints. This field will also be included, if present, in the standard JSON messages delivered to other endpoints. Constraints: Subjects must be ASCII text that begins with a letter, number, or punctuation mark; must not include line breaks or control characters; and must be less than 100 characters long.

_piTopicArn :: Maybe Text

The topic you want to publish to.

piTopicArn :: Lens' Publish (Maybe Text)

piTargetArn :: Lens' Publish (Maybe Text)

piSubject :: Lens' Publish (Maybe Text)

piMessageStructure :: Lens' Publish (Maybe Text)

piMessageAttributes :: Lens' Publish (Map Text MessageAttributeValue)

piMessage :: Lens' Publish Text

data PublishResponse

Constructors

PublishResponse

Fields

_prMessageId :: Maybe Text

Unique identifier assigned to the published message. Length Constraint: Maximum 100 characters.

prMessageId :: Iso' PublishResponse (Maybe Text)

data RemovePermission

Constructors

RemovePermission

Fields

_rpiLabel :: Text

The unique label of the statement you want to remove.

_rpiTopicArn :: Text

The ARN of the topic whose access control policy you wish to modify.

rpiTopicArn :: Lens' RemovePermission Text

rpiLabel :: Lens' RemovePermission Text

data RemovePermissionResponse

Constructors

RemovePermissionResponse

data SetEndpointAttributes

Constructors

SetEndpointAttributes

Fields

_seaiAttributes :: Map Text Text

A map of the endpoint attributes. Attributes in this map include the following: CustomUserData -- arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB. Enabled -- flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token. Token -- device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

_seaiEndpointArn :: Text

EndpointArn used for SetEndpointAttributes action.

seaiEndpointArn :: Lens' SetEndpointAttributes Text

seaiAttributes :: Lens' SetEndpointAttributes (Map Text Text)

data SetEndpointAttributesResponse

Constructors

SetEndpointAttributesResponse

data SetPlatformApplicationAttributes

Constructors

SetPlatformApplicationAttributes

Fields

_spaaiAttributes :: Map Text Text

A map of the platform application attributes. Attributes in this map include the following: PlatformCredential -- The credential received from the notification service. For APNS/APNS_SANDBOX, PlatformCredential is "private key". For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". PlatformPrincipal -- The principal received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is "SSL certificate". For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

_spaaiPlatformApplicationArn :: Text

PlatformApplicationArn for SetPlatformApplicationAttributes action.

spaaiPlatformApplicationArn :: Lens' SetPlatformApplicationAttributes Text

spaaiAttributes :: Lens' SetPlatformApplicationAttributes (Map Text Text)

data SetPlatformApplicationAttributesResponse

Constructors

SetPlatformApplicationAttributesResponse

setSubscriptionAttributes

Arguments

:: Text
_ssaiAttributeName

-> Text
_ssaiSubscriptionArn

-> SetSubscriptionAttributes

Minimum specification for a SetSubscriptionAttributes request.

data SetSubscriptionAttributes

Constructors

SetSubscriptionAttributes

Fields

_ssaiAttributeName :: Text

The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable. Valid values: DeliveryPolicy | RawMessageDelivery.

_ssaiSubscriptionArn :: Text

The ARN of the subscription to modify.

_ssaiAttributeValue :: Maybe Text

The new value for the attribute in JSON format.

ssaiSubscriptionArn :: Lens' SetSubscriptionAttributes Text

ssaiAttributeValue :: Lens' SetSubscriptionAttributes (Maybe Text)

ssaiAttributeName :: Lens' SetSubscriptionAttributes Text

data SetSubscriptionAttributesResponse

Constructors

SetSubscriptionAttributesResponse

setTopicAttributes

Arguments

:: Text
_staiAttributeName

-> Text
_staiTopicArn

-> SetTopicAttributes

Minimum specification for a SetTopicAttributes request.

data SetTopicAttributes

Constructors

SetTopicAttributes

Fields

_staiAttributeName :: Text

The name of the attribute you want to set. Only a subset of the topic's attributes are mutable. Valid values: Policy | DisplayName | DeliveryPolicy.

_staiTopicArn :: Text

The ARN of the topic to modify.

_staiAttributeValue :: Maybe Text

The new value for the attribute.

staiTopicArn :: Lens' SetTopicAttributes Text

staiAttributeValue :: Lens' SetTopicAttributes (Maybe Text)

staiAttributeName :: Lens' SetTopicAttributes Text

data SetTopicAttributesResponse

Constructors

SetTopicAttributesResponse

subscribe

Arguments

:: Text
_ssuProtocol

-> Text
_ssuTopicArn

-> Subscribe

Minimum specification for a Subscribe request.

data Subscribe

Constructors

Subscribe

Fields

_ssuProtocol :: Text

The protocol you want to use. Supported protocols include: http -- delivery of JSON-encoded message via HTTP POST https -- delivery of JSON-encoded message via HTTPS POST email -- delivery of message via SMTP email-json -- delivery of JSON-encoded message via SMTP sms -- delivery of message via SMS sqs -- delivery of JSON-encoded message to an Amazon SQS queue application -- delivery of JSON-encoded message to an EndpointArn for a mobile app and device.

_ssuTopicArn :: Text

The ARN of the topic you want to subscribe to.

_ssuEndpoint :: Maybe Text

The endpoint that you want to receive notifications. Endpoints vary by protocol: For the http protocol, the endpoint is an URL beginning with "http://" For the https protocol, the endpoint is a URL beginning with "https://" For the email protocol, the endpoint is an email address For the email-json protocol, the endpoint is an email address For the sms protocol, the endpoint is a phone number of an SMS-enabled device For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

ssuTopicArn :: Lens' Subscribe Text

ssuProtocol :: Lens' Subscribe Text

ssuEndpoint :: Lens' Subscribe (Maybe Text)

data SubscribeResponse

Constructors

SubscribeResponse

Fields

_sseSubscriptionArn :: Maybe Text

The ARN of the subscription, if the service was able to create a subscription immediately (without requiring endpoint owner confirmation).

sseSubscriptionArn :: Iso' SubscribeResponse (Maybe Text)

data SNS

Supported version (2010-03-31) of the Amazon Simple Notification Service service.

xmlOptions :: Tagged a XMLOptions

newtype Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes.

Constructors

Topic

Fields

_tTopicArn :: Maybe Text

The topic's ARN.

data Endpoint

Endpoint for mobile app and device.

Constructors

Endpoint

Fields

_fAttributes :: Map Text Text

Attributes for endpoint.

_fEndpointArn :: Maybe Text

EndpointArn for mobile app and device.

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish. Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Using Amazon SNS Message Attributes.

Constructors

MessageAttributeValue

Fields

_mavBinaryValue :: Maybe ByteString

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.

_mavStringValue :: Maybe Text

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters.

_mavDataType :: Text

Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types.

data PlatformApplication

Platform application object.

Constructors

PlatformApplication

Fields

_paPlatformApplicationArn :: Maybe Text

PlatformApplicationArn for platform application object.

_paAttributes :: Map Text Text

Attributes for platform application object.

data Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

Constructors

Subscription

Fields

_snProtocol :: Maybe Text

The subscription's protocol.

_snOwner :: Maybe Text

The subscription's owner.

_snTopicArn :: Maybe Text

The ARN of the subscription's topic.

_snEndpoint :: Maybe Text

The subscription's endpoint (format depends on the protocol).

_snSubscriptionArn :: Maybe Text

The subscription's ARN.

tTopicArn :: Iso' Topic (Maybe Text)

fEndpointArn :: Lens' Endpoint (Maybe Text)

fAttributes :: Lens' Endpoint (Map Text Text)

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

mavDataType :: Lens' MessageAttributeValue Text

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

paPlatformApplicationArn :: Lens' PlatformApplication (Maybe Text)

paAttributes :: Lens' PlatformApplication (Map Text Text)

snTopicArn :: Lens' Subscription (Maybe Text)

snSubscriptionArn :: Lens' Subscription (Maybe Text)

snProtocol :: Lens' Subscription (Maybe Text)

snOwner :: Lens' Subscription (Maybe Text)

snEndpoint :: Lens' Subscription (Maybe Text)

data Unsubscribe

Constructors

Unsubscribe

Fields

_uiSubscriptionArn :: Text

The ARN of the subscription to be deleted.

uiSubscriptionArn :: Iso' Unsubscribe Text

data UnsubscribeResponse

Constructors

UnsubscribeResponse