Network.AWS.SNS.V2010_03_31.AddPermission

Description

Adds a statement to a topic's access control policy, granting access for the specified AWS accounts to the specified actions. http:/sns.us-east-1.amazonaws.com ?TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Test &amp;ActionName.member.1=Publish &amp;ActionName.member.2=GetTopicAttributes &amp;Label=NewPermission &amp;AWSAccountId.member.1=987654321000 &amp;AWSAccountId.member.2=876543210000 &amp;Action=AddPermission &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=k%2FAU%2FKp13pjndwJ7rr1sZszy6MZMlOhRBCHx1ZaZFiw%3D &lt;AddPermissionResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;6a213e4e-33a8-11df-9540-99d0768312d3&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;AddPermissionResponse&gt;.

Request

data AddPermission

Request constructor

mkAddPermission

Arguments

:: Text
apTopicArn

-> Text
apLabel

-> [Text]
apAWSAccountId

-> [Text]
apActionName

-> AddPermission

Smart constructor for the minimum required parameters to construct a valid AddPermission request.

Request lenses

apTopicArn :: Lens' AddPermission Text

The ARN of the topic whose access control policy you wish to modify.

apLabel :: Lens' AddPermission Text

A unique identifier for the new policy statement.

apAWSAccountId :: Lens' AddPermission [Text]

The AWS account IDs of the users (principals) who will be given access to the specified actions. The users must have AWS accounts, but do not need to be signed up for this service.

apActionName :: Lens' AddPermission [Text]

The action you want to allow for the specified principal(s). Valid values: any Amazon SNS action name.

Response

data AddPermissionResponse

Response constructor

mkAddPermissionResponse :: AddPermissionResponse

Smart constructor for the minimum required parameters to construct a valid AddPermissionResponse response.

This constructor is provided for convenience and testing purposes.