Network.AWS.SNS.V2010_03_31.GetPlatformApplicationAttributes

Description

Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications. POST http:/sns.us-west-2.amazonaws.com HTTP/1.1 ... PlatformApplicationArn=arn%3Aaws%3Asns%3Aus-west-2%3A123456789012%3Aapp%2FGCM%2Fgcmpushapp &amp;Action=GetPlatformApplicationAttributes &amp;SignatureMethod=HmacSHA256 &amp;AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &amp;SignatureVersion=2 &amp;Version=2010-03-31 &amp;Signature=UGMaCq8CCJGSYXO9Ehr2VuHIBYSe6WbxkqgMKRslTK4%3D &amp;Timestamp=2013-07-01T22%3A40%3A50.643Z HTTP/1.1 200 OK ... &lt;GetPlatformApplicationAttributesResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;GetPlatformApplicationAttributesResult&gt; &lt;Attributes&gt; &lt;entry&gt; &lt;key&gt;AllowEndpointPolicies&lt;/key&gt; &lt;value&gt;false&lt;value&gt; &lt;entry&gt; &lt;/Attributes&gt; &lt;/GetPlatformApplicationAttributesResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;74848df2-87f6-55ed-890c-c7be80442462&lt;/RequestId&gt; &lt;/ResponseMetadata&gt; &lt;/GetPlatformApplicationAttributesResponse&gt;.

Request

data GetPlatformApplicationAttributes

Input for GetPlatformApplicationAttributes action.

Request constructor

mkGetPlatformApplicationAttributes

Arguments

:: Text
gpaaPlatformApplicationArn

-> GetPlatformApplicationAttributes

Smart constructor for the minimum required parameters to construct a valid GetPlatformApplicationAttributes request.

Request lenses

gpaaPlatformApplicationArn :: Lens' GetPlatformApplicationAttributes Text

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

Response

data GetPlatformApplicationAttributesResponse

Response for GetPlatformApplicationAttributes action.

Response lenses

gpaarAttributes :: Lens' GetPlatformApplicationAttributesResponse (Map Text Text)

Attributes include the following: EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.