Network.AWS.SNS.V2010_03_31.ListEndpointsByPlatformApplication

Description

Lists the endpoints and endpoint attributes for devices in a supported push notification service, such as GCM and APNS. The results for ListEndpointsByPlatformApplication are paginated and return a limited list of endpoints, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListEndpointsByPlatformApplication again using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications. POST http:/sns.us-west-2.amazonaws.com HTTP/1.1 ... PlatformApplicationArn=arn%3Aaws%3Asns%3Aus-west-2%3A123456789012%3Aapp%2FGCM%2Fgcmpushapp &amp;Action=ListEndpointsByPlatformApplication &amp;SignatureMethod=HmacSHA256 &amp;AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &amp;SignatureVersion=2 &amp;Version=2010-03-31 &amp;Signature=e6H4sJSCRBBlh%2BaigB%2FtYgp4%2Bjl7dikAQ6WKf%2BMTwNM%3D &amp;Timestamp=2013-07-01T23%3A00%3A52.515Z HTTP/1.1 200 OK ... &lt;ListEndpointsByPlatformApplicationResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ListEndpointsByPlatformApplicationResult&gt; &lt;Endpoints&gt; &lt;member&gt; &lt;EndpointArn&gt;arn:aws:sns:us-west-2:123456789012:endpointGCMgcmpushapp5e3e9847-3183-3f18-a7e8-671c3a57d4b3&lt;EndpointArn&gt; &lt;Attributes&gt; &lt;entry&gt; &lt;key&gt;Enabled&lt;/key&gt; &lt;value&gt;true&lt;value&gt; &lt;entry&gt; &lt;entry&gt; &lt;key&gt;CustomUserData&lt;/key&gt; &lt;value&gt;UserId=27576823&lt;value&gt; &lt;entry&gt; &lt;entry&gt; &lt;key&gt;Token&lt;/key&gt; &lt;value&gt;APA91bGi7fFachkC1xjlqT66VYEucGHochmf1VQAr9k...jsM0PKPxKhddCzx6paEsyay9Zn3D4wNUJb8m6HZrBEXAMPLE&lt;/value&gt; &lt;entry&gt; &lt;Attributes&gt; &lt;member&gt; &lt;Endpoints&gt; &lt;/ListEndpointsByPlatformApplicationResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;9a48768c-dac8-5a60-aec0-3cc27ea08d96&lt;/RequestId&gt; &lt;/ResponseMetadata&gt; &lt;/ListEndpointsByPlatformApplicationResponse&gt;.

Documentation

listEndpointsByPlatformApplication

Arguments

:: Text
_lebpaiPlatformApplicationArn

-> ListEndpointsByPlatformApplication

Minimum specification for a ListEndpointsByPlatformApplication request.

data ListEndpointsByPlatformApplication

Constructors

ListEndpointsByPlatformApplication

Fields

_lebpaiPlatformApplicationArn :: Text

PlatformApplicationArn for ListEndpointsByPlatformApplicationInput action.

_lebpaiNextToken :: Maybe Text

NextToken string is used when calling ListEndpointsByPlatformApplication action to retrieve additional records that are available after the first page results.

lebpaiPlatformApplicationArn :: Lens' ListEndpointsByPlatformApplication Text

lebpaiNextToken :: Lens' ListEndpointsByPlatformApplication (Maybe Text)

data ListEndpointsByPlatformApplicationResponse

Constructors

ListEndpointsByPlatformApplicationResponse

Fields

_lebparEndpoints :: [Endpoint]

Endpoints returned for ListEndpointsByPlatformApplication action.

_lebparNextToken :: Maybe Text

NextToken string is returned when calling ListEndpointsByPlatformApplication action if additional records are available after the first page results.

lebparNextToken :: Lens' ListEndpointsByPlatformApplicationResponse (Maybe Text)

lebparEndpoints :: Lens' ListEndpointsByPlatformApplicationResponse [Endpoint]