Network.AWS.SNS.V2010_03_31.ListPlatformApplications

Description

Lists the platform application objects for the supported push notification services, such as APNS and GCM. The results for ListPlatformApplications are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListPlatformApplications using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications. POST http:/sns.us-west-2.amazonaws.com HTTP/1.1 ... Action=ListPlatformApplications &amp;SignatureMethod=HmacSHA256 &amp;AWSAccessKeyId=AKIAIOSFODNN7EXAMPLE &amp;SignatureVersion=2 &amp;Version=2010-03-31 &amp;Signature=drVbTuyR5N9e88WJMNPzBOjNFNvawkCaMfZI0xa9kIQ%3D &amp;Timestamp=2013-07-01T22%3A33%3A55.618Z HTTP/1.1 200 OK ... &lt;ListPlatformApplicationsResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ListPlatformApplicationsResult&gt; &lt;PlatformApplications&gt; &lt;member&gt; &lt;PlatformApplicationArn&gt;arn:aws:sns:us-west-2:123456789012:appAPNS_SANDBOXapnspushapp&lt;/PlatformApplicationArn&gt; &lt;Attributes&gt; &lt;entry&gt; &lt;key&gt;AllowEndpointPolicies&lt;/key&gt; &lt;value&gt;false&lt;value&gt; &lt;entry&gt; &lt;/Attributes&gt; &lt;/member&gt; &lt;member&gt; &lt;PlatformApplicationArn&gt;arn:aws:sns:us-west-2:123456789012:appGCMgcmpushapp&lt;/PlatformApplicationArn&gt; &lt;Attributes&gt; &lt;entry&gt; &lt;key&gt;AllowEndpointPolicies&lt;/key&gt; &lt;value&gt;false&lt;value&gt; &lt;entry&gt; &lt;/Attributes&gt; &lt;member&gt; &lt;PlatformApplications&gt; &lt;/ListPlatformApplicationsResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;315a335e-85d8-52df-9349-791283cbb529&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;ListPlatformApplicationsResponse&gt;.

Documentation

listPlatformApplications :: ListPlatformApplications

Minimum specification for a ListPlatformApplications request.

data ListPlatformApplications

Constructors

ListPlatformApplications

Fields

_lpaiNextToken :: Maybe Text

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

lpaiNextToken :: Iso' ListPlatformApplications (Maybe Text)

data ListPlatformApplicationsResponse

Constructors

ListPlatformApplicationsResponse

Fields

_lparPlatformApplications :: [PlatformApplication]

Platform applications returned when calling ListPlatformApplications action.

_lparNextToken :: Maybe Text

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

lparPlatformApplications :: Lens' ListPlatformApplicationsResponse [PlatformApplication]

lparNextToken :: Lens' ListPlatformApplicationsResponse (Maybe Text)