Network.AWS.SNS.V2010_03_31.ListSubscriptions

Description

Returns a list of the requester's subscriptions. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptions call to get further results. http:/sns.us-east-1.amazonaws.com &amp;Action=ListSubscriptions &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=SZmBxEPqfs9R7xxhSt6C1b7PnOEvg%2BSVyyMYJfLRFCA%3D &lt;ListSubscriptionsResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ListSubscriptionsResult&gt; &lt;Subscriptions&gt; &lt;member&gt; &lt;TopicArn&gt;arn:aws:sns:us-east-1:698519295917:My-Topic&lt;/TopicArn&gt; &lt;Protocol&gt;email&lt;/Protocol&gt; &lt;SubscriptionArn&gt;arn:aws:sns:us-east-1:123456789012:My-Topic:80289ba6-0fd4-4079-afb4-ce8c8260f0ca&lt;/SubscriptionArn&gt; &lt;Owner&gt;123456789012&lt;/Owner&gt; &lt;Endpoint&gt;example@amazon.com&lt;Endpoint&gt; &lt;member&gt; &lt;Subscriptions&gt; &lt;ListSubscriptionsResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;384ac68d-3775-11df-8963-01868b7c937a&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;ListSubscriptionsResponse&gt;.

Request

data ListSubscriptions

Input for ListSubscriptions action.

Request constructor

mkListSubscriptions :: ListSubscriptions

Smart constructor for the minimum required parameters to construct a valid ListSubscriptions request.

Request lenses

lsNextToken :: Lens' ListSubscriptions (Maybe Text)

Token returned by the previous ListSubscriptions request.

Response

data ListSubscriptionsResponse

Response for ListSubscriptions action.

Response lenses

lsrSubscriptions :: Lens' ListSubscriptionsResponse [Subscription]

A list of subscriptions.

lsrNextToken :: Lens' ListSubscriptionsResponse (Maybe Text)

Token to pass along to the next ListSubscriptions request. This element is returned if there are more subscriptions to retrieve.