Network.AWS.SNS.V2010_03_31.ListSubscriptionsByTopic

Description

Returns a list of the subscriptions to a specific topic. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptionsByTopic call to get further results. http:/sns.us-east-1.amazonaws.com ?TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Topic &amp;Action=ListSubscriptionsByTopic &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=SZmBxEPqfs9R7xxhSt6C1b7PnOEvg%2BSVyyMYJfLRFCA%3D &lt;ListSubscriptionsByTopicResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ListSubscriptionsByTopicResult&gt; &lt;Subscriptions&gt; &lt;member&gt; &lt;TopicArn&gt;arn:aws:sns:us-east-1:123456789012:My-Topic&lt;/TopicArn&gt; &lt;Protocol&gt;email&lt;/Protocol&gt; &lt;SubscriptionArn&gt;arn:aws:sns:us-east-1:123456789012:My-Topic:80289ba6-0fd4-4079-afb4-ce8c8260f0ca&lt;/SubscriptionArn&gt; &lt;Owner&gt;123456789012&lt;/Owner&gt; &lt;Endpoint&gt;example@amazon.com&lt;Endpoint&gt; &lt;member&gt; &lt;Subscriptions&gt; &lt;ListSubscriptionsByTopicResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;b9275252-3774-11df-9540-99d0768312d3&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;ListSubscriptionsByTopicResponse&gt;.

Request

data ListSubscriptionsByTopic

Input for ListSubscriptionsByTopic action.

Request constructor

mkListSubscriptionsByTopic

Arguments

:: Text
lsbtTopicArn

-> ListSubscriptionsByTopic

Smart constructor for the minimum required parameters to construct a valid ListSubscriptionsByTopic request.

Request lenses

lsbtTopicArn :: Lens' ListSubscriptionsByTopic Text

The ARN of the topic for which you wish to find subscriptions.

lsbtNextToken :: Lens' ListSubscriptionsByTopic (Maybe Text)

Token returned by the previous ListSubscriptionsByTopic request.

Response

data ListSubscriptionsByTopicResponse

Response for ListSubscriptionsByTopic action.

Response lenses

lsbtrSubscriptions :: Lens' ListSubscriptionsByTopicResponse [Subscription]

A list of subscriptions.

lsbtrNextToken :: Lens' ListSubscriptionsByTopicResponse (Maybe Text)

Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.