Network.AWS.SNS.V2010_03_31.RemovePermission

Description

Removes a statement from a topic's access control policy. http:/sns.us-east-1.amazonaws.com ?TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Test &amp;Label=NewPermission &amp;Action=RemovePermission &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=N1abwRY9i7zaSQmbAlm71pPf9EEFOqNbQL1alzw2yCg%3D &lt;RemovePermissionResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;d170b150-33a8-11df-995a-2d6fbe836cc1&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;RemovePermissionResponse&gt;.

Documentation

data RemovePermission

Constructors

RemovePermission

Fields

_rpiLabel :: Text

The unique label of the statement you want to remove.

_rpiTopicArn :: Text

The ARN of the topic whose access control policy you wish to modify.

rpiTopicArn :: Lens' RemovePermission Text

rpiLabel :: Lens' RemovePermission Text

data RemovePermissionResponse

Constructors

RemovePermissionResponse