Network.AWS.SNS.V2010_03_31.SetTopicAttributes

Description

Allows a topic owner to set an attribute of the topic to a new value. The following example sets the DisplayName attribute to MyTopicName http:/sns.us-east-1.amazonaws.com ?AttributeValue=MyTopicName &amp;TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Topic &amp;AttributeName=DisplayName &amp;Action=SetTopicAttributes &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key Id) &amp;Signature=mQA3nJI%2BcmAIY7r8HCArGElSqPX5JG4UGzF4yo0RygE%3D The following example sets the delivery policy to 5 total retries http:/sns.us-east-1.amazonaws.com ?AttributeValue={"http":{"defaultHealthyRetryPolicy":{"numRetries":5}}} &amp;TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Topic &amp;AttributeName=DeliveryPolicy &amp;Action=SetTopicAttributes &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key Id) &amp;Signature=mQA3nJI%2BcmAIY7r8HCArGElSqPX5JG4UGzF4yo0RygE%3D The JSON format for the DeliveryPolicy AttributeValue (linebreaks added for readability): { "http": { "defaultHealthyRetryPolicy": { "minDelayTarget": &lt;int&gt;, "maxDelayTarget": &lt;int&gt;, "numRetries": &lt;int&gt;, "numMaxDelayRetries": &lt;int&gt;, "backoffFunction": "&lt;linear|arithmetic|geometric|exponential&gt;" }, "disableSubscriptionOverrides": &lt;boolean&gt;, "defaultThrottlePolicy": { "maxReceivesPerSecond": &lt;int&gt; } } &lt;SetTopicAttributesResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;a8763b99-33a7-11df-a9b7-05d48da6f042&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;SetTopicAttributesResponse&gt;.

Documentation

setTopicAttributes

Arguments

:: Text
_staiAttributeName

-> Text
_staiTopicArn

-> SetTopicAttributes

Minimum specification for a SetTopicAttributes request.

data SetTopicAttributes

Constructors

SetTopicAttributes

Fields

_staiAttributeName :: Text

The name of the attribute you want to set. Only a subset of the topic's attributes are mutable. Valid values: Policy | DisplayName | DeliveryPolicy.

_staiTopicArn :: Text

The ARN of the topic to modify.

_staiAttributeValue :: Maybe Text

The new value for the attribute.

staiTopicArn :: Lens' SetTopicAttributes Text

staiAttributeValue :: Lens' SetTopicAttributes (Maybe Text)

staiAttributeName :: Lens' SetTopicAttributes Text

data SetTopicAttributesResponse

Constructors

SetTopicAttributesResponse