Network.AWS.SNS.V2010_03_31.Subscribe

Description

Prepares to subscribe an endpoint by sending the endpoint a confirmation message. To actually create a subscription, the endpoint owner must call the ConfirmSubscription action with the token from the confirmation message. Confirmation tokens are valid for three days. The following example Query request subscribes an SQS queue to an SNS topic. For more information, see Subscribe Queue to Amazon SNS Topic in the Amazon SQS Developer Guide. http:/sns.us-west-2.amazonaws.com &amp;Action=Subscribe &amp;Endpoint=arn%3Aaws%3Asqs%3Aus-west-2%3A123456789012%3AMyQueue &amp;Version=2010-03-31 &amp;Protocol=sqs &amp;TopicArn=arn%3Aaws%3Asns%3Aus-west-2%3A123456789012%3AMyTopic &lt;SubscribeResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;SubscribeResult&gt; &lt;SubscriptionArn&gt;arn:aws:sns:us-west-2:123456789012:MyTopic:6b0e71bd-7e97-4d97-80ce-4a0994e55286&lt;/SubscriptionArn&gt; &lt;/SubscribeResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;c4407779-24a4-56fa-982c-3d927f93a775&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;SubscribeResponse&gt;.

Request

data Subscribe

Input for Subscribe action.

Request constructor

mkSubscribe

Arguments

:: Text
s1TopicArn

-> Text
s1Protocol

-> Subscribe

Smart constructor for the minimum required parameters to construct a valid Subscribe request.

Request lenses

s1TopicArn :: Lens' Subscribe Text

The ARN of the topic you want to subscribe to.

s1Protocol :: Lens' Subscribe Text

The protocol you want to use. Supported protocols include: http -- delivery of JSON-encoded message via HTTP POST https -- delivery of JSON-encoded message via HTTPS POST email -- delivery of message via SMTP email-json -- delivery of JSON-encoded message via SMTP sms -- delivery of message via SMS sqs -- delivery of JSON-encoded message to an Amazon SQS queue application -- delivery of JSON-encoded message to an EndpointArn for a mobile app and device.

s1Endpoint :: Lens' Subscribe (Maybe Text)

The endpoint that you want to receive notifications. Endpoints vary by protocol: For the http protocol, the endpoint is an URL beginning with "http://" For the https protocol, the endpoint is a URL beginning with "https://" For the email protocol, the endpoint is an email address For the email-json protocol, the endpoint is an email address For the sms protocol, the endpoint is a phone number of an SMS-enabled device For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

Response

data SubscribeResponse

Response for Subscribe action.

Response constructor

mkSubscribeResponse :: SubscribeResponse

Smart constructor for the minimum required parameters to construct a valid SubscribeResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

srSubscriptionArn :: Lens' SubscribeResponse (Maybe Text)

The ARN of the subscription, if the service was able to create a subscription immediately (without requiring endpoint owner confirmation).