Network.AWS.SNS

Description

Amazon Simple Notification Service

Amazon Simple Notification Service (Amazon SNS) is a web service that enables you to build distributed web-enabled applications. Applications can use Amazon SNS to easily push real-time notification messages to interested subscribers over multiple delivery protocols. For more information about this product see http://aws.amazon.com/sns . For detailed information about Amazon SNS features and their associated API calls, see the Amazon SNS Developer Guide .

We also provide SDKs that enable you to access Amazon SNS from your preferred programming language. The SDKs contain functionality that automatically takes care of tasks such as: cryptographically signing your service requests, retrying requests, and handling error responses. For a list of available SDKs, go to Tools for Amazon Web Services .

Service Configuration

sns :: Service

API version 2010-03-31 of the Amazon Simple Notification Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SNS.

EndpointDisabledException

_EndpointDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception error indicating endpoint disabled.

AuthorizationErrorException

_AuthorizationErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the user has been denied access to the requested resource.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

SubscriptionLimitExceededException

_SubscriptionLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the customer already owns the maximum allowed number of subscriptions.

PlatformApplicationDisabledException

_PlatformApplicationDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception error indicating platform application disabled.

InternalErrorException

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an internal service error.

ThrottledException

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the rate at which requests have been submitted for this action exceeds the limit for your account.

InvalidParameterValueException

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the requested resource does not exist.

TopicLimitExceededException

_TopicLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the customer already owns the maximum allowed number of topics.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListPhoneNumbersOptedOut

module Network.AWS.SNS.ListPhoneNumbersOptedOut

DeleteEndpoint

module Network.AWS.SNS.DeleteEndpoint

RemovePermission

module Network.AWS.SNS.RemovePermission

DeleteTopic

module Network.AWS.SNS.DeleteTopic

SetSMSAttributes

module Network.AWS.SNS.SetSMSAttributes

ListTopics (Paginated)

module Network.AWS.SNS.ListTopics

CreatePlatformEndpoint

module Network.AWS.SNS.CreatePlatformEndpoint

SetPlatformApplicationAttributes

module Network.AWS.SNS.SetPlatformApplicationAttributes

ListSubscriptionsByTopic (Paginated)

module Network.AWS.SNS.ListSubscriptionsByTopic

GetTopicAttributes

module Network.AWS.SNS.GetTopicAttributes

OptInPhoneNumber

module Network.AWS.SNS.OptInPhoneNumber

CreatePlatformApplication

module Network.AWS.SNS.CreatePlatformApplication

GetPlatformApplicationAttributes

module Network.AWS.SNS.GetPlatformApplicationAttributes

ListEndpointsByPlatformApplication (Paginated)

module Network.AWS.SNS.ListEndpointsByPlatformApplication

SetTopicAttributes

module Network.AWS.SNS.SetTopicAttributes

DeletePlatformApplication

module Network.AWS.SNS.DeletePlatformApplication

GetSMSAttributes

module Network.AWS.SNS.GetSMSAttributes

ListPlatformApplications (Paginated)

module Network.AWS.SNS.ListPlatformApplications

AddPermission

module Network.AWS.SNS.AddPermission

GetEndpointAttributes

module Network.AWS.SNS.GetEndpointAttributes

ListSubscriptions (Paginated)

module Network.AWS.SNS.ListSubscriptions

GetSubscriptionAttributes

module Network.AWS.SNS.GetSubscriptionAttributes

CreateTopic

module Network.AWS.SNS.CreateTopic

CheckIfPhoneNumberIsOptedOut

module Network.AWS.SNS.CheckIfPhoneNumberIsOptedOut

Subscribe

module Network.AWS.SNS.Subscribe

Unsubscribe

module Network.AWS.SNS.Unsubscribe

SetEndpointAttributes

module Network.AWS.SNS.SetEndpointAttributes

SetSubscriptionAttributes

module Network.AWS.SNS.SetSubscriptionAttributes

ConfirmSubscription

module Network.AWS.SNS.ConfirmSubscription

Publish

module Network.AWS.SNS.Publish

Types

Endpoint

data Endpoint

Endpoint for mobile app and device.

See: endpoint smart constructor.

endpoint :: Endpoint

Creates a value of Endpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eAttributes - Attributes for endpoint.

eEndpointARN - EndpointArn for mobile app and device.

eAttributes :: Lens' Endpoint (HashMap Text Text)

Attributes for endpoint.

eEndpointARN :: Lens' Endpoint (Maybe Text)

EndpointArn for mobile app and device.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish .

Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Using Amazon SNS Message Attributes .

See: messageAttributeValue smart constructor.

messageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

Creates a value of MessageAttributeValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mavBinaryValue - Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

mavStringValue - Strings are Unicode with UTF8 binary encoding. For a list of code values, see http://en.wikipedia.org/wiki/ASCII#ASCII_printable_characters .

mavDataType - Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types .

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http://en.wikipedia.org/wiki/ASCII#ASCII_printable_characters .

mavDataType :: Lens' MessageAttributeValue Text

Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types .

PlatformApplication

data PlatformApplication

Platform application object.

See: platformApplication smart constructor.

platformApplication :: PlatformApplication

Creates a value of PlatformApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paPlatformApplicationARN - PlatformApplicationArn for platform application object.

paAttributes - Attributes for platform application object.

paPlatformApplicationARN :: Lens' PlatformApplication (Maybe Text)

PlatformApplicationArn for platform application object.

paAttributes :: Lens' PlatformApplication (HashMap Text Text)

Attributes for platform application object.

Subscription

data Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

See: subscription smart constructor.

subscription :: Subscription

Creates a value of Subscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sProtocol - The subscription's protocol.

sOwner - The subscription's owner.

sTopicARN - The ARN of the subscription's topic.

sEndpoint - The subscription's endpoint (format depends on the protocol).

sSubscriptionARN - The subscription's ARN.

sProtocol :: Lens' Subscription (Maybe Text)

The subscription's protocol.

sOwner :: Lens' Subscription (Maybe Text)

The subscription's owner.

sTopicARN :: Lens' Subscription (Maybe Text)

The ARN of the subscription's topic.

sEndpoint :: Lens' Subscription (Maybe Text)

The subscription's endpoint (format depends on the protocol).

sSubscriptionARN :: Lens' Subscription (Maybe Text)

The subscription's ARN.

Topic

data Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes .

See: topic smart constructor.

topic :: Topic

Creates a value of Topic with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tTopicARN - The topic's ARN.

tTopicARN :: Lens' Topic (Maybe Text)

The topic's ARN.