Network.AWS.SNS.ConfirmSubscription

Description

Verifies an endpoint owner's intent to receive messages by validating the token sent to the endpoint by an earlier Subscribe action. If the token is valid, the action creates a new subscription and returns its Amazon Resource Name (ARN). This call requires an AWS signature only when the AuthenticateOnUnsubscribe flag is set to "true".

Creating a Request

confirmSubscription

Arguments

:: Text
csTopicARN

-> Text
csToken

-> ConfirmSubscription

Creates a value of ConfirmSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csAuthenticateOnUnsubscribe - Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an AWS signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires AWS authentication.

csTopicARN - The ARN of the topic for which you wish to confirm a subscription.

csToken - Short-lived token sent to an endpoint during the Subscribe action.

data ConfirmSubscription

Input for ConfirmSubscription action.

See: confirmSubscription smart constructor.

Request Lenses

csAuthenticateOnUnsubscribe :: Lens' ConfirmSubscription (Maybe Text)

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an AWS signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires AWS authentication.

csTopicARN :: Lens' ConfirmSubscription Text

The ARN of the topic for which you wish to confirm a subscription.

csToken :: Lens' ConfirmSubscription Text

Short-lived token sent to an endpoint during the Subscribe action.

Destructuring the Response

confirmSubscriptionResponse

Arguments

:: Int
csrsResponseStatus

-> ConfirmSubscriptionResponse

Creates a value of ConfirmSubscriptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csrsSubscriptionARN - The ARN of the created subscription.

csrsResponseStatus - -- | The response status code.

data ConfirmSubscriptionResponse

Response for ConfirmSubscriptions action.

See: confirmSubscriptionResponse smart constructor.

Response Lenses

csrsSubscriptionARN :: Lens' ConfirmSubscriptionResponse (Maybe Text)

The ARN of the created subscription.

csrsResponseStatus :: Lens' ConfirmSubscriptionResponse Int

- | The response status code.