Network.AWS.SNS.CreatePlatformApplication

Description

Creates a platform application object for one of the supported push notification services, such as APNS and GCM, to which devices and mobile apps may register. You must specify PlatformPrincipal and PlatformCredential attributes when using the CreatePlatformApplication action. The PlatformPrincipal is received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is "SSL certificate". For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". The PlatformCredential is also received from the notification service. For WNS, PlatformPrincipal is "Package Security Identifier". For MPNS, PlatformPrincipal is "TLS certificate". For Baidu, PlatformPrincipal is "API key".

For APNSAPNS_SANDBOX, PlatformCredential is "private key". For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". For WNS, PlatformCredential is "secret key". For MPNS, PlatformCredential is "private key". For Baidu, PlatformCredential is "secret key". The PlatformApplicationArn that is returned when using CreatePlatformApplication is then used as an attribute for the CreatePlatformEndpoint action. For more information, see <http:docs.aws.amazon.comsnslatestdgSNSMobilePush.html Using Amazon SNS Mobile Push Notifications> . For more information about obtaining the PlatformPrincipal and PlatformCredential for each of the supported push notification services, see <http:docs.aws.amazon.comsnslatestdgmobile-push-apns.html Getting Started with Apple Push Notification Service> , <http:docs.aws.amazon.comsnslatestdgmobile-push-adm.html Getting Started with Amazon Device Messaging> , <http:docs.aws.amazon.comsnslatestdgmobile-push-baidu.html Getting Started with Baidu Cloud Push> , <http:docs.aws.amazon.comsnslatestdgmobile-push-gcm.html Getting Started with Google Cloud Messaging for Android> , <http:docs.aws.amazon.comsnslatestdgmobile-push-mpns.html Getting Started with MPNS> , or <http:docs.aws.amazon.comsnslatestdg/mobile-push-wns.html Getting Started with WNS> .

Creating a Request

createPlatformApplication

Arguments

:: Text
cpaName

-> Text
cpaPlatform

-> CreatePlatformApplication

Creates a value of CreatePlatformApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpaName - Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

cpaPlatform - The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Google Cloud Messaging).

cpaAttributes - For a list of attributes, see SetPlatformApplicationAttributes

data CreatePlatformApplication

Input for CreatePlatformApplication action.

See: createPlatformApplication smart constructor.

Request Lenses

cpaName :: Lens' CreatePlatformApplication Text

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

cpaPlatform :: Lens' CreatePlatformApplication Text

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Google Cloud Messaging).

cpaAttributes :: Lens' CreatePlatformApplication (HashMap Text Text)

For a list of attributes, see SetPlatformApplicationAttributes

Destructuring the Response

createPlatformApplicationResponse

Arguments

:: Int
cparsResponseStatus

-> CreatePlatformApplicationResponse

Creates a value of CreatePlatformApplicationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cparsPlatformApplicationARN - PlatformApplicationArn is returned.

cparsResponseStatus - -- | The response status code.

data CreatePlatformApplicationResponse

Response from CreatePlatformApplication action.

See: createPlatformApplicationResponse smart constructor.

Response Lenses

cparsPlatformApplicationARN :: Lens' CreatePlatformApplicationResponse (Maybe Text)

PlatformApplicationArn is returned.

cparsResponseStatus :: Lens' CreatePlatformApplicationResponse Int

- | The response status code.