Network.AWS.SNS.CreatePlatformEndpoint

Description

Creates an endpoint for a device and mobile app on one of the supported push notification services, such as GCM and APNS. CreatePlatformEndpoint requires the PlatformApplicationArn that is returned from CreatePlatformApplication . The EndpointArn that is returned when using CreatePlatformEndpoint can then be used by the Publish action to send a message to a mobile app or by the Subscribe action for subscription to a topic. The CreatePlatformEndpoint action is idempotent, so if the requester already owns an endpoint with the same device token and attributes, that endpoint's ARN is returned without creating a new endpoint. For more information, see Using Amazon SNS Mobile Push Notifications .

When using CreatePlatformEndpoint with Baidu, two attributes must be provided: ChannelId and UserId. The token field must also contain the ChannelId. For more information, see Creating an Amazon SNS Endpoint for Baidu .

Creating a Request

createPlatformEndpoint

Arguments

:: Text
cpePlatformApplicationARN

-> Text
cpeToken

-> CreatePlatformEndpoint

Creates a value of CreatePlatformEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpeCustomUserData - Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

cpeAttributes - For a list of attributes, see SetEndpointAttributes .

cpePlatformApplicationARN - PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

cpeToken - Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM or ADM, the device token equivalent is called the registration ID.

data CreatePlatformEndpoint

Input for CreatePlatformEndpoint action.

See: createPlatformEndpoint smart constructor.

Request Lenses

cpeCustomUserData :: Lens' CreatePlatformEndpoint (Maybe Text)

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

cpeAttributes :: Lens' CreatePlatformEndpoint (HashMap Text Text)

For a list of attributes, see SetEndpointAttributes .

cpePlatformApplicationARN :: Lens' CreatePlatformEndpoint Text

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

cpeToken :: Lens' CreatePlatformEndpoint Text

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM or ADM, the device token equivalent is called the registration ID.

Destructuring the Response

createPlatformEndpointResponse

Arguments

:: Int
cpersResponseStatus

-> CreatePlatformEndpointResponse

Creates a value of CreatePlatformEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpersEndpointARN - EndpointArn returned from CreateEndpoint action.

cpersResponseStatus - -- | The response status code.

data CreatePlatformEndpointResponse

Response from CreateEndpoint action.

See: createPlatformEndpointResponse smart constructor.

Response Lenses

cpersEndpointARN :: Lens' CreatePlatformEndpointResponse (Maybe Text)

EndpointArn returned from CreateEndpoint action.

cpersResponseStatus :: Lens' CreatePlatformEndpointResponse Int

- | The response status code.