Network.AWS.SNS.CreatePlatformEndpoint

Description

Creates an endpoint for a device and mobile app on one of the supported push notification services, such as GCM and APNS. CreatePlatformEndpoint requires the PlatformApplicationArn that is returned from CreatePlatformApplication. The EndpointArn that is returned when using CreatePlatformEndpoint can then be used by the Publish action to send a message to a mobile app or by the Subscribe action for subscription to a topic. The CreatePlatformEndpoint action is idempotent, so if the requester already owns an endpoint with the same device token and attributes, that endpoint's ARN is returned without creating a new endpoint. For more information, see Using Amazon SNS Mobile Push Notifications .

When using CreatePlatformEndpoint with Baidu, two attributes must be provided: ChannelId and UserId. The token field must also contain the ChannelId. For more information, see Creating an Amazon SNS Endpoint for Baidu .

http://docs.aws.amazon.com/sns/latest/api/API_CreatePlatformEndpoint.html

Request

data CreatePlatformEndpoint

Request constructor

createPlatformEndpoint

Arguments

:: Text
cpePlatformApplicationArn

-> Text
cpeToken

-> CreatePlatformEndpoint

CreatePlatformEndpoint constructor.

The fields accessible through corresponding lenses are:

cpeAttributes :: HashMap Text Text

cpeCustomUserData :: Maybe Text

cpePlatformApplicationArn :: Text

cpeToken :: Text

Request lenses

cpeAttributes :: Lens' CreatePlatformEndpoint (HashMap Text Text)

For a list of attributes, see SetEndpointAttributes.

cpeCustomUserData :: Lens' CreatePlatformEndpoint (Maybe Text)

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

cpePlatformApplicationArn :: Lens' CreatePlatformEndpoint Text

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

cpeToken :: Lens' CreatePlatformEndpoint Text

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM or ADM, the device token equivalent is called the registration ID.

Response

data CreatePlatformEndpointResponse

Response constructor

createPlatformEndpointResponse :: CreatePlatformEndpointResponse

CreatePlatformEndpointResponse constructor.

The fields accessible through corresponding lenses are:

cperEndpointArn :: Maybe Text

Response lenses

cperEndpointArn :: Lens' CreatePlatformEndpointResponse (Maybe Text)

EndpointArn returned from CreateEndpoint action.