Network.AWS.SNS.CreateTopic

Description

Creates a topic to which notifications can be published. Users can create at most 100,000 topics. For more information, see http://aws.amazon.com/sns . This action is idempotent, so if the requester already owns a topic with the specified name, that topic's ARN is returned without creating a new topic.

Creating a Request

createTopic

Arguments

:: Text
ctName

-> CreateTopic

Creates a value of CreateTopic with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctName - The name of the topic you want to create. Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

data CreateTopic

Input for CreateTopic action.

See: createTopic smart constructor.

Request Lenses

ctName :: Lens' CreateTopic Text

The name of the topic you want to create. Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

Destructuring the Response

createTopicResponse

Arguments

:: Int
ctrsResponseStatus

-> CreateTopicResponse

Creates a value of CreateTopicResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrsTopicARN - The Amazon Resource Name (ARN) assigned to the created topic.

ctrsResponseStatus - -- | The response status code.

data CreateTopicResponse

Response from CreateTopic action.

See: createTopicResponse smart constructor.

Response Lenses

ctrsTopicARN :: Lens' CreateTopicResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the created topic.

ctrsResponseStatus :: Lens' CreateTopicResponse Int

- | The response status code.