Network.AWS.SNS.DeleteTopic

Description

Deletes a topic and all its subscriptions. Deleting a topic might prevent some messages previously sent to the topic from being delivered to subscribers. This action is idempotent, so deleting a topic that does not exist does not result in an error.

http://docs.aws.amazon.com/sns/latest/api/API_DeleteTopic.html

Request

data DeleteTopic

Request constructor

deleteTopic

Arguments

:: Text
dtTopicArn

-> DeleteTopic

DeleteTopic constructor.

The fields accessible through corresponding lenses are:

dtTopicArn :: Text

Request lenses

dtTopicArn :: Lens' DeleteTopic Text

The ARN of the topic you want to delete.

Response

data DeleteTopicResponse

Response constructor

deleteTopicResponse :: DeleteTopicResponse

DeleteTopicResponse constructor.