Network.AWS.SNS.GetEndpointAttributes

Description

Retrieves the endpoint attributes for a device on one of the supported push notification services, such as GCM and APNS. For more information, see Using Amazon SNS Mobile Push Notifications .

Creating a Request

getEndpointAttributes

Arguments

:: Text
geaEndpointARN

-> GetEndpointAttributes

Creates a value of GetEndpointAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

geaEndpointARN - EndpointArn for GetEndpointAttributes input.

data GetEndpointAttributes

Input for GetEndpointAttributes action.

See: getEndpointAttributes smart constructor.

Request Lenses

geaEndpointARN :: Lens' GetEndpointAttributes Text

EndpointArn for GetEndpointAttributes input.

Destructuring the Response

getEndpointAttributesResponse

Arguments

:: Int
gearsResponseStatus

-> GetEndpointAttributesResponse

Creates a value of GetEndpointAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gearsAttributes - Attributes include the following: * CustomUserData -- arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB. * Enabled -- flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token. * Token -- device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

gearsResponseStatus - -- | The response status code.

data GetEndpointAttributesResponse

Response from GetEndpointAttributes of the EndpointArn.

See: getEndpointAttributesResponse smart constructor.

Response Lenses

gearsAttributes :: Lens' GetEndpointAttributesResponse (HashMap Text Text)

Attributes include the following: * CustomUserData -- arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB. * Enabled -- flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token. * Token -- device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

gearsResponseStatus :: Lens' GetEndpointAttributesResponse Int

- | The response status code.