Network.AWS.SNS.GetPlatformApplicationAttributes

Description

Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications.

http://docs.aws.amazon.com/sns/latest/api/API_GetPlatformApplicationAttributes.html

Request

data GetPlatformApplicationAttributes

Request constructor

getPlatformApplicationAttributes

Arguments

:: Text
gpaaPlatformApplicationArn

-> GetPlatformApplicationAttributes

GetPlatformApplicationAttributes constructor.

The fields accessible through corresponding lenses are:

gpaaPlatformApplicationArn :: Text

Request lenses

gpaaPlatformApplicationArn :: Lens' GetPlatformApplicationAttributes Text

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

Response

data GetPlatformApplicationAttributesResponse

Response constructor

getPlatformApplicationAttributesResponse :: GetPlatformApplicationAttributesResponse

GetPlatformApplicationAttributesResponse constructor.

The fields accessible through corresponding lenses are:

gpaarAttributes :: HashMap Text Text

Response lenses

gpaarAttributes :: Lens' GetPlatformApplicationAttributesResponse (HashMap Text Text)

Attributes include the following:

EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.