Network.AWS.SNS.GetSubscriptionAttributes

Description

Returns all of the properties of a subscription.

Creating a Request

getSubscriptionAttributes

Arguments

:: Text
gsaSubscriptionARN

-> GetSubscriptionAttributes

Creates a value of GetSubscriptionAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsaSubscriptionARN - The ARN of the subscription whose properties you want to get.

data GetSubscriptionAttributes

Input for GetSubscriptionAttributes.

See: getSubscriptionAttributes smart constructor.

Request Lenses

gsaSubscriptionARN :: Lens' GetSubscriptionAttributes Text

The ARN of the subscription whose properties you want to get.

Destructuring the Response

getSubscriptionAttributesResponse

Arguments

:: Int
gsarsResponseStatus

-> GetSubscriptionAttributesResponse

Creates a value of GetSubscriptionAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsarsAttributes - A map of the subscription's attributes. Attributes in this map include the following: * SubscriptionArn -- the subscription's ARN * TopicArn -- the topic ARN that the subscription is associated with * Owner -- the AWS account ID of the subscription's owner * ConfirmationWasAuthenticated -- true if the subscription confirmation request was authenticated * DeliveryPolicy -- the JSON serialization of the subscription's delivery policy * EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account the topic delivery policy and account system defaults

gsarsResponseStatus - -- | The response status code.

data GetSubscriptionAttributesResponse

Response for GetSubscriptionAttributes action.

See: getSubscriptionAttributesResponse smart constructor.

Response Lenses

gsarsAttributes :: Lens' GetSubscriptionAttributesResponse (HashMap Text Text)

A map of the subscription's attributes. Attributes in this map include the following: * SubscriptionArn -- the subscription's ARN * TopicArn -- the topic ARN that the subscription is associated with * Owner -- the AWS account ID of the subscription's owner * ConfirmationWasAuthenticated -- true if the subscription confirmation request was authenticated * DeliveryPolicy -- the JSON serialization of the subscription's delivery policy * EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account the topic delivery policy and account system defaults

gsarsResponseStatus :: Lens' GetSubscriptionAttributesResponse Int

- | The response status code.