Network.AWS.SNS.GetTopicAttributes

Description

Returns all of the properties of a topic. Topic properties returned might differ based on the authorization of the user.

Creating a Request

getTopicAttributes

Arguments

:: Text
gtaTopicARN

-> GetTopicAttributes

Creates a value of GetTopicAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtaTopicARN - The ARN of the topic whose properties you want to get.

data GetTopicAttributes

Input for GetTopicAttributes action.

See: getTopicAttributes smart constructor.

Request Lenses

gtaTopicARN :: Lens' GetTopicAttributes Text

The ARN of the topic whose properties you want to get.

Destructuring the Response

getTopicAttributesResponse

Arguments

:: Int
gtarsResponseStatus

-> GetTopicAttributesResponse

Creates a value of GetTopicAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtarsAttributes - A map of the topic's attributes. Attributes in this map include the following: * TopicArn -- the topic's ARN * Owner -- the AWS account ID of the topic's owner * Policy -- the JSON serialization of the topic's access control policy * DisplayName -- the human-readable name used in the From field for notifications to email and email-json endpoints * SubscriptionsPending -- the number of subscriptions pending confirmation on this topic * SubscriptionsConfirmed -- the number of confirmed subscriptions on this topic * SubscriptionsDeleted -- the number of deleted subscriptions on this topic * DeliveryPolicy -- the JSON serialization of the topic's delivery policy * EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account system defaults

gtarsResponseStatus - -- | The response status code.

data GetTopicAttributesResponse

Response for GetTopicAttributes action.

See: getTopicAttributesResponse smart constructor.

Response Lenses

gtarsAttributes :: Lens' GetTopicAttributesResponse (HashMap Text Text)

A map of the topic's attributes. Attributes in this map include the following: * TopicArn -- the topic's ARN * Owner -- the AWS account ID of the topic's owner * Policy -- the JSON serialization of the topic's access control policy * DisplayName -- the human-readable name used in the From field for notifications to email and email-json endpoints * SubscriptionsPending -- the number of subscriptions pending confirmation on this topic * SubscriptionsConfirmed -- the number of confirmed subscriptions on this topic * SubscriptionsDeleted -- the number of deleted subscriptions on this topic * DeliveryPolicy -- the JSON serialization of the topic's delivery policy * EffectiveDeliveryPolicy -- the JSON serialization of the effective delivery policy that takes into account system defaults

gtarsResponseStatus :: Lens' GetTopicAttributesResponse Int

- | The response status code.