Network.AWS.SNS.ListPlatformApplications

Description

Lists the platform application objects for the supported push notification services, such as APNS and GCM. The results for ListPlatformApplications are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListPlatformApplications using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications .

This operation returns paginated results.

Creating a Request

listPlatformApplications :: ListPlatformApplications

Creates a value of ListPlatformApplications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpaNextToken - NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

data ListPlatformApplications

Input for ListPlatformApplications action.

See: listPlatformApplications smart constructor.

Request Lenses

lpaNextToken :: Lens' ListPlatformApplications (Maybe Text)

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

Destructuring the Response

listPlatformApplicationsResponse

Arguments

:: Int
lparsResponseStatus

-> ListPlatformApplicationsResponse

Creates a value of ListPlatformApplicationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lparsPlatformApplications - Platform applications returned when calling ListPlatformApplications action.

lparsNextToken - NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

lparsResponseStatus - -- | The response status code.

data ListPlatformApplicationsResponse

Response for ListPlatformApplications action.

See: listPlatformApplicationsResponse smart constructor.

Response Lenses

lparsPlatformApplications :: Lens' ListPlatformApplicationsResponse [PlatformApplication]

Platform applications returned when calling ListPlatformApplications action.

lparsNextToken :: Lens' ListPlatformApplicationsResponse (Maybe Text)

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

lparsResponseStatus :: Lens' ListPlatformApplicationsResponse Int

- | The response status code.