Network.AWS.SNS.ListSubscriptions

Description

Returns a list of the requester's subscriptions. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptions call to get further results.

This operation returns paginated results.

Creating a Request

listSubscriptions :: ListSubscriptions

Creates a value of ListSubscriptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsNextToken - Token returned by the previous ListSubscriptions request.

data ListSubscriptions

Input for ListSubscriptions action.

See: listSubscriptions smart constructor.

Request Lenses

lsNextToken :: Lens' ListSubscriptions (Maybe Text)

Token returned by the previous ListSubscriptions request.

Destructuring the Response

listSubscriptionsResponse

Arguments

:: Int
lsrsResponseStatus

-> ListSubscriptionsResponse

Creates a value of ListSubscriptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsrsNextToken - Token to pass along to the next ListSubscriptions request. This element is returned if there are more subscriptions to retrieve.

lsrsSubscriptions - A list of subscriptions.

lsrsResponseStatus - -- | The response status code.

data ListSubscriptionsResponse

Response for ListSubscriptions action

See: listSubscriptionsResponse smart constructor.

Response Lenses

lsrsNextToken :: Lens' ListSubscriptionsResponse (Maybe Text)

Token to pass along to the next ListSubscriptions request. This element is returned if there are more subscriptions to retrieve.

lsrsSubscriptions :: Lens' ListSubscriptionsResponse [Subscription]

A list of subscriptions.

lsrsResponseStatus :: Lens' ListSubscriptionsResponse Int

- | The response status code.