Network.AWS.SNS.Publish

Description

Sends a message to all of a topic's subscribed endpoints. When a messageId is returned, the message has been saved and Amazon SNS will attempt to deliver it to the topic's subscribers shortly. The format of the outgoing message to each subscribed endpoint depends on the notification protocol selected.

To use the Publish action for sending a message to a mobile endpoint, such as an app on a Kindle device or mobile phone, you must specify the EndpointArn. The EndpointArn is returned when making a call with the CreatePlatformEndpoint action. The second example below shows a request and response for publishing to a mobile endpoint.

http://docs.aws.amazon.com/sns/latest/api/API_Publish.html

Request

data Publish

Request constructor

publish

Arguments

:: Text
pMessage

-> Publish

Publish constructor.

The fields accessible through corresponding lenses are:

pMessage :: Text

pMessageAttributes :: HashMap Text MessageAttributeValue

pMessageStructure :: Maybe Text

pSubject :: Maybe Text

pTargetArn :: Maybe Text

pTopicArn :: Maybe Text

Request lenses

pMessage :: Lens' Publish Text

The message you want to send to the topic.

If you want to send the same message to all transport protocols, include the text of the message as a String value.

If you want to send different messages for each transport protocol, set the value of the MessageStructure parameter to json and use a JSON object for the Message parameter. See the Examples section for the format of the JSON object.

Constraints: Messages must be UTF-8 encoded strings at most 256 KB in size (262144 bytes, not 262144 characters).

JSON-specific constraints: Keys in the JSON object that correspond to supported transport protocols must have simple JSON string values. The values will be parsed (unescaped) before they are used in outgoing messages. Outbound notifications are JSON encoded (meaning that the characters will be reescaped for sending). Values have a minimum length of 0 (the empty string, "", is allowed). Values have a maximum length bounded by the overall message size (so, including multiple protocols may limit message sizes). Non-string values will cause the key to be ignored. Keys that do not correspond to supported transport protocols are ignored. Duplicate keys are not allowed. Failure to parse or validate any key or value in the message will cause the Publish call to return an error (no partial delivery).

pMessageAttributes :: Lens' Publish (HashMap Text MessageAttributeValue)

Message attributes for Publish action.

pMessageStructure :: Lens' Publish (Maybe Text)

Set MessageStructure to json if you want to send a different message for each protocol. For example, using one publish action, you can send a short message to your SMS subscribers and a longer message to your email subscribers. If you set MessageStructure to json, the value of the Message parameter must:

be a syntactically valid JSON object; and contain at least a top-level JSON key of "default" with a value that is a string. You can define other top-level keys that define the message you want to send to a specific transport protocol (e.g., "http").

For information about sending different messages for each protocol using the AWS Management Console, go to Create Different Messages for Each Protocol in the Amazon Simple Notification Service Getting Started Guide.

Valid value: json

pSubject :: Lens' Publish (Maybe Text)

Optional parameter to be used as the Subject line when the message is delivered to email endpoints. This field will also be included, if present, in the standard JSON messages delivered to other endpoints.

Constraints: Subjects must be ASCII text that begins with a letter, number, or punctuation mark; must not include line breaks or control characters; and must be less than 100 characters long.

pTargetArn :: Lens' Publish (Maybe Text)

Either TopicArn or EndpointArn, but not both.

pTopicArn :: Lens' Publish (Maybe Text)

The topic you want to publish to.

Response

data PublishResponse

Response constructor

publishResponse :: PublishResponse

PublishResponse constructor.

The fields accessible through corresponding lenses are:

prMessageId :: Maybe Text

Response lenses

prMessageId :: Lens' PublishResponse (Maybe Text)

Unique identifier assigned to the published message.

Length Constraint: Maximum 100 characters