Network.AWS.SNS.SetPlatformApplicationAttributes

Description

Sets the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications . For information on configuring attributes for message delivery status, see Using Amazon SNS Application Attributes for Message Delivery Status .

Creating a Request

setPlatformApplicationAttributes

Arguments

:: Text
spaaPlatformApplicationARN

-> SetPlatformApplicationAttributes

Creates a value of SetPlatformApplicationAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spaaPlatformApplicationARN - PlatformApplicationArn for SetPlatformApplicationAttributes action.

spaaAttributes - A map of the platform application attributes. Attributes in this map include the following: * PlatformCredential -- The credential received from the notification service. For APNSAPNS_SANDBOX, PlatformCredential is private key. For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". * PlatformPrincipal -- The principal received from the notification service. For APNSAPNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". * EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. * EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. * EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. * EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints. * SuccessFeedbackRoleArn -- IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf. * FailureFeedbackRoleArn -- IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf. * SuccessFeedbackSampleRate -- Sample rate percentage (0-100) of successfully delivered messages.

data SetPlatformApplicationAttributes

Input for SetPlatformApplicationAttributes action.

See: setPlatformApplicationAttributes smart constructor.

Request Lenses

spaaPlatformApplicationARN :: Lens' SetPlatformApplicationAttributes Text

PlatformApplicationArn for SetPlatformApplicationAttributes action.

spaaAttributes :: Lens' SetPlatformApplicationAttributes (HashMap Text Text)

A map of the platform application attributes. Attributes in this map include the following: * PlatformCredential -- The credential received from the notification service. For APNSAPNS_SANDBOX, PlatformCredential is private key. For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". * PlatformPrincipal -- The principal received from the notification service. For APNSAPNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". * EventEndpointCreated -- Topic ARN to which EndpointCreated event notifications should be sent. * EventEndpointDeleted -- Topic ARN to which EndpointDeleted event notifications should be sent. * EventEndpointUpdated -- Topic ARN to which EndpointUpdate event notifications should be sent. * EventDeliveryFailure -- Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints. * SuccessFeedbackRoleArn -- IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf. * FailureFeedbackRoleArn -- IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf. * SuccessFeedbackSampleRate -- Sample rate percentage (0-100) of successfully delivered messages.

Destructuring the Response

setPlatformApplicationAttributesResponse :: SetPlatformApplicationAttributesResponse

Creates a value of SetPlatformApplicationAttributesResponse with the minimum fields required to make a request.

data SetPlatformApplicationAttributesResponse

See: setPlatformApplicationAttributesResponse smart constructor.