Network.AWS.SNS.SetSubscriptionAttributes

Description

Allows a subscription owner to set an attribute of the topic to a new value.

Creating a Request

setSubscriptionAttributes

Arguments

:: Text
ssaSubscriptionARN

-> Text
ssaAttributeName

-> SetSubscriptionAttributes

Creates a value of SetSubscriptionAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssaAttributeValue - The new value for the attribute in JSON format.

ssaSubscriptionARN - The ARN of the subscription to modify.

ssaAttributeName - The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable. Valid values: DeliveryPolicy | RawMessageDelivery

data SetSubscriptionAttributes

Input for SetSubscriptionAttributes action.

See: setSubscriptionAttributes smart constructor.

Request Lenses

ssaAttributeValue :: Lens' SetSubscriptionAttributes (Maybe Text)

The new value for the attribute in JSON format.

ssaSubscriptionARN :: Lens' SetSubscriptionAttributes Text

The ARN of the subscription to modify.

ssaAttributeName :: Lens' SetSubscriptionAttributes Text

The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable. Valid values: DeliveryPolicy | RawMessageDelivery

Destructuring the Response

setSubscriptionAttributesResponse :: SetSubscriptionAttributesResponse

Creates a value of SetSubscriptionAttributesResponse with the minimum fields required to make a request.

data SetSubscriptionAttributesResponse

See: setSubscriptionAttributesResponse smart constructor.