Network.AWS.SNS.SetSubscriptionAttributes

Description

Allows a subscription owner to set an attribute of the topic to a new value.

http://docs.aws.amazon.com/sns/latest/api/API_SetSubscriptionAttributes.html

Request

data SetSubscriptionAttributes

Request constructor

setSubscriptionAttributes

Arguments

:: Text
ssaSubscriptionArn

-> Text
ssaAttributeName

-> SetSubscriptionAttributes

SetSubscriptionAttributes constructor.

The fields accessible through corresponding lenses are:

ssaAttributeName :: Text

ssaAttributeValue :: Maybe Text

ssaSubscriptionArn :: Text

Request lenses

ssaAttributeName :: Lens' SetSubscriptionAttributes Text

The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable.

Valid values: DeliveryPolicy | RawMessageDelivery

ssaAttributeValue :: Lens' SetSubscriptionAttributes (Maybe Text)

The new value for the attribute in JSON format.

ssaSubscriptionArn :: Lens' SetSubscriptionAttributes Text

The ARN of the subscription to modify.

Response

data SetSubscriptionAttributesResponse

Response constructor

setSubscriptionAttributesResponse :: SetSubscriptionAttributesResponse

SetSubscriptionAttributesResponse constructor.