Network.AWS.SNS.Types

Description

Service Configuration

sns :: Service

API version 2010-03-31 of the Amazon Simple Notification Service SDK configuration.

Errors

_EndpointDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception error indicating endpoint disabled.

_AuthorizationErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the user has been denied access to the requested resource.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

_SubscriptionLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the customer already owns the maximum allowed number of subscriptions.

_PlatformApplicationDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception error indicating platform application disabled.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an internal service error.

_ThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the rate at which requests have been submitted for this action exceeds the limit for your account.

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request parameter does not comply with the associated constraints.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the requested resource does not exist.

_TopicLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the customer already owns the maximum allowed number of topics.

Endpoint

data Endpoint

Endpoint for mobile app and device.

See: endpoint smart constructor.

endpoint :: Endpoint

Creates a value of Endpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eAttributes - Attributes for endpoint.

eEndpointARN - EndpointArn for mobile app and device.

eAttributes :: Lens' Endpoint (HashMap Text Text)

Attributes for endpoint.

eEndpointARN :: Lens' Endpoint (Maybe Text)

EndpointArn for mobile app and device.

MessageAttributeValue

data MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute has the same restrictions on the content as the message body. For more information, see Publish .

Name, type, and value must not be empty or null. In addition, the message body should not be empty or null. All parts of the message attribute, including name, type, and value, are included in the message size restriction, which is currently 256 KB (262,144 bytes). For more information, see Using Amazon SNS Message Attributes .

See: messageAttributeValue smart constructor.

messageAttributeValue

Arguments

:: Text
mavDataType

-> MessageAttributeValue

Creates a value of MessageAttributeValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mavBinaryValue - Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

mavStringValue - Strings are Unicode with UTF8 binary encoding. For a list of code values, see http://en.wikipedia.org/wiki/ASCII#ASCII_printable_characters .

mavDataType - Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types .

mavBinaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

mavStringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see http://en.wikipedia.org/wiki/ASCII#ASCII_printable_characters .

mavDataType :: Lens' MessageAttributeValue Text

Amazon SNS supports the following logical data types: String, Number, and Binary. For more information, see Message Attribute Data Types .

PlatformApplication

data PlatformApplication

Platform application object.

See: platformApplication smart constructor.

platformApplication :: PlatformApplication

Creates a value of PlatformApplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paPlatformApplicationARN - PlatformApplicationArn for platform application object.

paAttributes - Attributes for platform application object.

paPlatformApplicationARN :: Lens' PlatformApplication (Maybe Text)

PlatformApplicationArn for platform application object.

paAttributes :: Lens' PlatformApplication (HashMap Text Text)

Attributes for platform application object.

Subscription

data Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

See: subscription smart constructor.

subscription :: Subscription

Creates a value of Subscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sProtocol - The subscription's protocol.

sOwner - The subscription's owner.

sTopicARN - The ARN of the subscription's topic.

sEndpoint - The subscription's endpoint (format depends on the protocol).

sSubscriptionARN - The subscription's ARN.

sProtocol :: Lens' Subscription (Maybe Text)

The subscription's protocol.

sOwner :: Lens' Subscription (Maybe Text)

The subscription's owner.

sTopicARN :: Lens' Subscription (Maybe Text)

The ARN of the subscription's topic.

sEndpoint :: Lens' Subscription (Maybe Text)

The subscription's endpoint (format depends on the protocol).

sSubscriptionARN :: Lens' Subscription (Maybe Text)

The subscription's ARN.

Topic

data Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's attributes, use GetTopicAttributes .

See: topic smart constructor.

topic :: Topic

Creates a value of Topic with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tTopicARN - The topic's ARN.

tTopicARN :: Lens' Topic (Maybe Text)

The topic's ARN.