Network.AWS.SNS.V2010_03_31.ListTopics

Description

Returns a list of the requester's topics. Each call returns a limited list of topics, up to 100. If there are more topics, a NextToken is also returned. Use the NextToken parameter in a new ListTopics call to get further results. http:/sns.us-east-1.amazonaws.com ?Action=ListTopics &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=tPg1qKNTNVPydnL3Yx5Fqm2O9GxCr9vh3EF5r9%2F5%2BJs%3D &lt;ListTopicsResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ListTopicsResult&gt; &lt;Topics&gt; &lt;member&gt; &lt;TopicArn&gt;arn:aws:sns:us-east-1:123456789012:My-Topic&lt;/TopicArn&gt; &lt;member&gt; &lt;Topics&gt; &lt;/ListTopicsResult&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;3f1478c7-33a9-11df-9540-99d0768312d3&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;ListTopicsResponse&gt;.

Documentation

listTopics :: ListTopics

Minimum specification for a ListTopics request.

data ListTopics

Constructors

ListTopics

Fields

_ltiNextToken :: Maybe Text

Token returned by the previous ListTopics request.

ltiNextToken :: Iso' ListTopics (Maybe Text)

data ListTopicsResponse

Constructors

ListTopicsResponse

Fields

_ltrNextToken :: Maybe Text

Token to pass along to the next ListTopics request. This element is returned if there are additional topics to retrieve.

_ltrTopics :: [Topic]

A list of topic ARNs.

ltrTopics :: Lens' ListTopicsResponse [Topic]

ltrNextToken :: Lens' ListTopicsResponse (Maybe Text)