Network.AWS.SNS.V2010_03_31.RemovePermission

Description

Removes a statement from a topic's access control policy. http:/sns.us-east-1.amazonaws.com ?TopicArn=arn%3Aaws%3Asns%3Aus-east-1%3A123456789012%3AMy-Test &amp;Label=NewPermission &amp;Action=RemovePermission &amp;SignatureVersion=2 &amp;SignatureMethod=HmacSHA256 &amp;Timestamp=2010-03-31T12%3A00%3A00.000Z &amp;AWSAccessKeyId=(AWS Access Key ID) &amp;Signature=N1abwRY9i7zaSQmbAlm71pPf9EEFOqNbQL1alzw2yCg%3D &lt;RemovePermissionResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/"&gt; &lt;ResponseMetadata&gt; &lt;RequestId&gt;d170b150-33a8-11df-995a-2d6fbe836cc1&lt;/RequestId&gt; &lt;ResponseMetadata&gt; &lt;RemovePermissionResponse&gt;.

Request

data RemovePermission

Input for RemovePermission action.

Request constructor

mkRemovePermission

Arguments

:: Text
rpTopicArn

-> Text
rpLabel

-> RemovePermission

Smart constructor for the minimum required parameters to construct a valid RemovePermission request.

Request lenses

rpTopicArn :: Lens' RemovePermission Text

The ARN of the topic whose access control policy you wish to modify.

rpLabel :: Lens' RemovePermission Text

The unique label of the statement you want to remove.

Response

data RemovePermissionResponse

Response constructor

mkRemovePermissionResponse :: RemovePermissionResponse

Smart constructor for the minimum required parameters to construct a valid RemovePermissionResponse response.

This constructor is provided for convenience and testing purposes.